% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTSmh.R
\name{convert_base4frac_to_dec}
\alias{convert_base4frac_to_dec}
\title{Convert a vector of base 4 fractions into (truncated) decimal integer
values}
\usage{
convert_base4frac_to_dec(x, level)
}
\arguments{
\item{x}{A scalar or vector of base 4 fractions, originating from the
\code{GRTSmaster_habitats} data source.}

\item{level}{The number of leading digits to discard from the GRTS base 4
address, i.e. from the '\code{xxx...}' digits behind the decimal mark in
the '\code{0.xxxxxxxxxxxxx}' base 4 fractions.
Only values from \code{0} (maintain full address) to \code{12} (only return
last digit) are sensible, as the GRTS addresses are \code{13} digits long.}
}
\value{
The corresponding decimal (i.e. base 10) integer scalar or vector.
}
\description{
Converts base 4 fractions, representing the full GRTS addresses from the
raster data source \code{GRTSmaster_habitats} into decimal (i.e. base 10)
integer values.
Before the actual conversion happens, leading digits from the full GRTS
address can be discarded according to the \code{level} specified by the user.
Hence, the result may correspond to the GRTS ranking at a lower spatial
resolution.
}
\details{
For example, the base 4 fraction \code{0.0000000000100}
is converted into decimal integer \code{16} (\code{= 4^2}) as long as the
\code{level} argument is \code{10} or lower (if not, it will be \code{0}) and
 \code{0.0000000000101} is converted into either \code{17} (
 \code{level <= 10}) or \code{1} (if \code{level} is \code{11} or \code{12}).

Long base 4 fractions seem to be handled and stored easier than long
(base 4) integers.
This approach follows the one of Stevens & Olsen (2004) to represent
the reverse hierarchical order in a GRTS sample.

The function works on a vector and retains \code{NA} values.
As such, it can be used in \code{raster::calc()}.
When writing such a raster to a file, it is recommended to use the
\code{INT4U} data type (see \code{\link[raster]{dataType}}).
}
\examples{
oldoption <- options(list(digits = 15, scipen = 999))
# one scalar:
convert_base4frac_to_dec(0.1010101010101, level = 0)
# vector, level 0:
convert_base4frac_to_dec(c(NA, 0.1010101010101), level = 0)
# vector, level 5:
convert_base4frac_to_dec(c(NA, 0.1010101010101), level = 5)
# same vector, all sensible levels computed:
sapply(0:12, function(i) {
  convert_base4frac_to_dec(c(NA, 0.1010101010101),
    level = i
  )
})
options(oldoption)

}
\references{
Stevens D.L. & Olsen A.R. (2004). Spatially Balanced Sampling of Natural
Resources. Journal of the American Statistical Association 99 (465):
262–278. \doi{10.1198/016214504000000250}.
}
\seealso{
Other functions involved in processing the 'GRTSmaster_habitats' data source:
\code{\link[=convert_dec_to_base4frac]{convert_dec_to_base4frac()}},
\code{\link[=read_GRTSmh]{read_GRTSmh()}},
\code{\link[=read_GRTSmh_base4frac]{read_GRTSmh_base4frac()}},
\code{\link[=read_GRTSmh_diffres]{read_GRTSmh_diffres()}}
}
\concept{functions involved in processing the 'GRTSmaster_habitats' data source}
