% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Get data from a Zenodo archive}
\usage{
download_zenodo(doi, path = ".", parallel = TRUE, quiet = FALSE)
}
\arguments{
\item{doi}{a doi pointer to the Zenodo archive starting with '10.5281/zenodo.'. See examples.}

\item{path}{Path where the data must be downloaded.
Defaults to the working directory.}

\item{parallel}{Logical.
If \code{TRUE} (the default), files will be
downloaded concurrently for multi-file records.
Of course, the operation is limited by bandwidth and traffic limitations.}

\item{quiet}{Logical (\code{FALSE} by default).
Do you want to suppress informative messages (not warnings)?}
}
\description{
This function will download an entire archive from Zenodo (\url{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)
}
\examples{
\dontrun{
# Example download of an archive containing a single zip
download_zenodo(doi = "10.5281/zenodo.1283345")
download_zenodo(doi = "10.5281/zenodo.1283345", quiet = TRUE)
# Example download of an archive containing multiple files
# using parallel download
# (multiple files will be simultaneously downloaded)
download_zenodo(doi = "10.5281/zenodo.1172801", parallel = TRUE)
# Example download of an archive containing a single pdf file
download_zenodo(doi = "10.5281/zenodo.168478")
}
}
\seealso{
Other functions regarding file management for N2KHAB projects:
\code{\link[=checksum]{checksum()}},
\code{\link[=convert_line_endings]{convert_line_endings()}},
\code{\link[=fileman_folders]{fileman_folders()}},
\code{\link[=fileman_up]{fileman_up()}},
\code{\link[=get_zenodo_versions]{get_zenodo_versions()}},
\code{\link[=locate_n2khab_data]{locate_n2khab_data()}}
}
\author{
Hans Van Calster, \email{hans.vancalster@inbo.be}

Floris Vanderhaeghe, \email{floris.vanderhaeghe@inbo.be}
}
\concept{functions regarding file management for N2KHAB projects}
