% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrangling.R
\name{expand_types_plain}
\alias{expand_types_plain}
\title{Translates a factor with levels mapped according to the namelist}
\usage{
expand_types_plain(x, type_var = "type", strict = TRUE)
}
\arguments{
\item{x}{An object of class \code{data.frame}.}

\item{type_var}{A string.
The name of the dataframe variable that holds the type codes.
Defaults to \code{type}.}

\item{strict}{Logical.
Apply conditions before expanding subtype codes to main type codes?}
}
\value{
A dataframe.
}
\description{
Translates a factor, by default using the 'name' variable from the
'namelist' object to substitute the factor levels. The join on 'namelist' is
by the variable 'code'.
}
\keyword{internal}
