% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{fileman_folders}
\alias{fileman_folders}
\title{Create a standard data folder structure and return the path to the \code{n2khab_data} folder}
\usage{
fileman_folders(root = c("rproj", "git"), path = NA)
}
\arguments{
\item{root}{Character string indicating whether the root folder of the current git repository or the root folder of the current Rstudio project should be used as the folder where you want the data folder structure to be created.
Can be \code{"rproj"} (the default) for an RStudio R project or \code{"git"}
for a git repository.}

\item{path}{An optional argument to specify a custom path to a folder where you want the data folder structure to be created. Default is \code{NA} (no custom path).}
}
\value{
A character string that gives the absolute path to the \code{n2khab_data/}
folder.
}
\description{
This function will check for the existence of default data folders,
create them if necessary, and return the path to the \code{n2khab_data} folder.
}
\details{
In n2khab projects a standardized folder setup is used for binary
data, as explained in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
The functions creates the folders \code{n2khab_data}, \code{n2khab_data/10_raw}
and \code{n2khab_data/20_processed}, or prints a message if these already
exist.
The function returns the path to \code{n2khab_data}.
}
\examples{
\dontrun{
fileman_folders()
datapath <- fileman_folders(root = "git")
}

}
\seealso{
Other functions regarding file management for N2KHAB projects:
\code{\link[=checksum]{checksum()}},
\code{\link[=convert_line_endings]{convert_line_endings()}},
\code{\link[=download_zenodo]{download_zenodo()}},
\code{\link[=fileman_up]{fileman_up()}},
\code{\link[=get_zenodo_versions]{get_zenodo_versions()}},
\code{\link[=locate_n2khab_data]{locate_n2khab_data()}}
}
\concept{functions regarding file management for N2KHAB projects}
