% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{fileman_zenodo}
\alias{fileman_zenodo}
\title{Get data from a Zenodo archive}
\usage{
fileman_zenodo(path, doi)
}
\arguments{
\item{path}{local path to where the data need to be written}

\item{doi}{a doi pointer to the Zenodo archive starting with '10.5281/zenodo.'. See examples.}
}
\value{
Downloaded file(s) in the specified folder.
}
\description{
This function will download data from Zenodo (\url{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)
}
\examples{
\dontrun{
# Single zip file deposition
fileman_zenodo(".", "10.5281/zenodo.1283345")
# Multiple files deposition
fileman_zenodo(".", "10.5281/zenodo.1172801")
# Single pdf file deposition
fileman_zenodo(".", "10.5281/zenodo.168478")
}
}
