% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{filemanag_folders}
\alias{filemanag_folders}
\title{Create a standard data folder structure and return the path to the \code{data} folder}
\usage{
filemanag_folders(root = c("rproj", "git"), path = NA)
}
\arguments{
\item{root}{Character string indicating whether the root folder of the current git repository or the root folder of the current Rstudio project should be used as the folder where you want the data folder structure to be created.
Can be \code{"rproj"} (the default) for an RStudio R project or \code{"git"}
for a git repository.}

\item{path}{An optional argument to specify a custom path to a folder where you want the data folder structure to be created. Default is \code{NA} (no custom path).}
}
\value{
A character string that gives the absolute path to the \code{data/}
folder.
}
\description{
This function will check for the existence of default data folders,
create them if necessary, and return the path to the data folder.
}
\details{
In n2khab projects a standardized folder setup is used for binary
data, as explained in the \href{doc/vign-020_datastorage.html}{vignette} on
data storage.
The functions creates the folders \code{data}, \code{data/10_raw}
and \code{data/20_processed}, or prints a message if these already
exist.
You can use the value returned by the function as the `path` argument of
functions that read particular data.
}
\examples{
filemanag_folders()
datapath <- filemanag_folders(root = "git")

}
