% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{filemanag_processed}
\alias{filemanag_processed}
\title{Get processed data}
\usage{
filemanag_processed(path, filename)
}
\arguments{
\item{path}{path to where the data need to be written}

\item{filename}{character vector of filenames}
}
\value{
downloaded files in the specified folder
}
\description{
This function will download processed data so they become locally available
}
\keyword{internal}
