% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{filemanag_zenodo}
\alias{filemanag_zenodo}
\title{Get raw data from a zenodo archive}
\usage{
filemanag_zenodo(path, doi)
}
\arguments{
\item{path}{path to where the data need to be written}

\item{doi}{doi (a pointer to the zenodo archive)}
}
\value{
downloaded and unzipped file in the folder
}
\description{
This function will download data from zenodo
}
\keyword{internal}
