% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{locate_n2khab_data}
\alias{locate_n2khab_data}
\title{Locate the \code{n2khab_data} directory}
\usage{
locate_n2khab_data()
}
\description{
Returns the absolute path of the \code{n2khab_data} directory by searching
upwards in the directory hierarchy (starting in the working directory); the
first hit will be returned.
}
\details{
If the \code{n2khab_data_path} option (or environment variable
\code{N2KHAB_DATA_PATH}) is set, that value will be returned instead (see
\code{\link[=n2khab_options]{n2khab_options()}}).

See the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")})
for more information.
}
\examples{
\dontrun{
locate_n2khab_data()
}

}
\seealso{
Other functions regarding file management for N2KHAB projects:
\code{\link[=checksum]{checksum()}},
\code{\link[=convert_line_endings]{convert_line_endings()}},
\code{\link[=download_zenodo]{download_zenodo()}},
\code{\link[=fileman_folders]{fileman_folders()}},
\code{\link[=fileman_up]{fileman_up()}},
\code{\link[=get_zenodo_versions]{get_zenodo_versions()}}
}
\concept{functions regarding file management for N2KHAB projects}
