% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{namelist_factor}
\alias{namelist_factor}
\title{Translates a factor with levels mapped according to the namelist}
\usage{
namelist_factor(x, pick = "name", codelist)
}
\arguments{
\item{x}{A factor.}

\item{pick}{A character, referring to the variable from namelist
(\code{"name"} or \code{"shortname"}) that should be used as translated
factor levels.}

\item{codelist}{An object conforming to \code{\link{namelist}}.}
}
\value{
A factor.
}
\description{
Translates a factor, by default using the 'name' variable from the
'namelist' object to substitute the factor levels. The join on 'namelist' is
by the variable 'code'.
}
\keyword{internal}
