% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{pkgdatasource_path}
\alias{pkgdatasource_path}
\title{Return the path of a package data source}
\usage{
pkgdatasource_path(file, extension = "")
}
\arguments{
\item{file}{The filename of the data source. Can be without extension.
Can be a character vector.}

\item{extension}{The extension of a file if not provided by \code{file}.
The dot is to be included, e.g. \code{.csv}.}
}
\value{
A character vector.
}
\description{
Returns the path for a given file
basename in a package. The developer can optionally
provide the file's extension separately.
}
\keyword{internal}
