% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecoregions.R
\name{read_ecoregions}
\alias{read_ecoregions}
\title{Return the \code{ecoregions} data source as an \code{sf} object}
\usage{
read_ecoregions(path = fileman_up("n2khab_data"), file = "10_raw/ecoregions")
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns the raw data source \code{ecoregions}, with unique polygon
identifiers \code{polygon_code} and \code{polygon_id}.
Multiple polygons can have the same \code{region_name}.
}
\details{
Original columns of the raw data source were mapped as:
\itemize{
\item \code{CODE} -> \code{polygon_code}
\item \code{NR} -> \code{polygon_id}
\item \code{REGIO} -> \code{region_name}
\item \code{DISTRICT} -> \code{district_name}
}

Apart from the label, there is no complementary information between
\code{polygon_code} and \code{polygon_id}.
}
\examples{
\dontrun{
ecoregions <- read_ecoregions()
ecoregions
}

}
