% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTSmh.R
\name{read_GRTSmh_diffres}
\alias{read_GRTSmh_diffres}
\title{Return a SpatRaster or an \code{sf} polygon layer from the processed data
source \code{GRTSmh_diffres}}
\usage{
read_GRTSmh_diffres(
  dir = file.path(locate_n2khab_data(), "20_processed/GRTSmh_diffres"),
  level,
  polygon = FALSE
)
}
\arguments{
\item{dir}{The data source directory (absolute or relative).
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{level}{Integer in the range from 1 to 9; determines the spatial
resolution. See the Details section.}

\item{polygon}{Logical; determines whether a polygon layer or a
SpatRaster is returned. See the Details section.}
}
\value{
Either a SpatRaster or a Simple feature collection of geometry type
\code{POLYGON}.

If the package is configured to use the raster package
(see \code{\link[=n2khab_options]{n2khab_options()}}), a RasterLayer is
returned instead of a SpatRaster.
}
\description{
The \code{GRTSmh_diffres} data source is derived from
\code{GRTSmh_brick}.
It provides the hierarchical levels 1 to 9 of the
GRTS cell addresses at the corresponding spatial resolution.
The function returns one selected level, either as a SpatRaster or as an
\code{sf} polygon layer (in the latter case, only levels 4 to 9 are
provided).
The coordinate reference system is 'BD72 / Belgian Lambert 72'
(EPSG-code \href{https://epsg.io/31370}{31370}).
}
\details{
The \code{GRTSmh_diffres} data source file is a file collection (available
at \href{https://doi.org/10.5281/zenodo.3354405}{Zenodo}), composed of
nine monolayered GeoTIFF files of the \code{INT4S} datatype plus a GeoPackage
with six polygon layers:
\itemize{
\item{The polygon layers in the GeoPackage are the dissolved, polygonized
versions of levels 4 to 9 of the
\code{GRTSmh_brick} data source (see \code{\link{read_GRTSmh}}).
This means that they provide the decimal (i.e. base 10) integer values of
these \emph{higher hierarchical levels} of the GRTS cell addresses
of the raw data source \code{GRTSmaster_habitats}.
Hence, the polygons are typically squares that correspond to the GRTS cell at
the specified hierarchical level.
The polygon layer is however restricted to the non-\code{NA} cells of the original
\code{GRTSmaster_habitats} raster.
Consequently, a part of the polygons is clipped along the Flemish border.
Levels 1 to 3 are not provided for the whole of Flanders,
because this would inflate the GPKG file.
You can look at the \href{https://github.com/inbo/n2khab-preprocessing}{source code}
to do such things.
}
\item{The GeoTIFF files provide the respective levels 1 to 9 of the
\code{GRTSmh_brick} data source in a raster format, at the resolution that
corresponds to the GRTS cell at the specified hierarchical level.
The presence of \code{NA} cells around Flanders at level 0 implies that, with
decreasing resolution, the raster's extent increases and larger areas outside
Flanders are covered by non-\code{NA} cells along the border.
}
}

The function returns the selected \code{level} either as an \code{sf} polygon
layer or as a SpatRaster, depending on the
\code{polygon} argument.

The higher-level ranking numbers (compared to the original level 0) allow
spatially balanced samples at lower spatial resolution than that of 32 m,
and can also be used for aggregation purposes.

The levels 1 to 9 correspond to the resolutions vector
\code{32 * 2^(1:9)} in meters:

\tabular{rr}{
level \tab    resolution (m)\cr
1 \tab    64\cr
2 \tab   128\cr
3 \tab   256\cr
4 \tab   512\cr
5 \tab  1024\cr
6 \tab  2048\cr
7 \tab  4096\cr
8 \tab  8192\cr
9 \tab 16384
}

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation from
the \code{GRTSmh_brick} data source.

Beware that not all GRTS ranking numbers at the specified level are
provided, as the original GRTS raster has been clipped with the Flemish
outer borders (i.e., not excluding the Brussels Capital Region).
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has
# the 'n2khab_data' folder AND that the latest version of the
# 'GRTSmh_diffres' data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can consider
# what to do.
r <- read_GRTSmh_diffres(level = 7)
r
p <- read_GRTSmh_diffres(level = 7, polygon = TRUE)
p
}

}
\seealso{
Other functions involved in processing the 'GRTSmaster_habitats' data source:
\code{\link[=convert_base4frac_to_dec]{convert_base4frac_to_dec()}},
\code{\link[=convert_dec_to_base4frac]{convert_dec_to_base4frac()}},
\code{\link[=read_GRTSmh]{read_GRTSmh()}},
\code{\link[=read_GRTSmh_base4frac]{read_GRTSmh_base4frac()}}
}
\concept{functions involved in processing the 'GRTSmaster_habitats' data source}
