% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatquarries}
\alias{read_habitatquarries}
\title{Return the data source \code{habitatquarries}}
\usage{
read_habitatquarries(
  file = file.path(fileman_up("n2khab_data"),
    "10_raw/habitatquarries/habitatquarries.gpkg"),
  filter_hab = FALSE,
  references = FALSE,
  bibtex = FALSE,
  version = "habitatquarries_2020v1"
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{filter_hab}{If \code{TRUE}, only polygons with (known) habitat \code{8310}
are returned.}

\item{references}{If \code{TRUE}, a list is returned with both the \code{sf}
object (element \code{habitatquarries}) and a dataframe of bibliographic
references (element \code{extra_references}).}

\item{bibtex}{If \code{TRUE}, all that happens is bibliographic references
being printed to the console, formatted for usage in a BibTeX file (\verb{*.bib}).}

\item{version}{Version ID of the data source.
Defaults to the latest available version defined by the package.}
}
\value{
Depending on the arguments, one of:
\itemize{
\item a simple feature collection of
type \code{POLYGON}, with attribute variables:
\itemize{
\item \code{polygon_id}: a unique number per polygon.
\item \code{unit_id}: a unique number for each quarry unit. Quarry units
consisting of several polygons (= partly outside the Flemish region)
have a number greater than 100.
\item \code{name}: site name.
\item \code{code_orig}: original 'habitattype' code in the raw data
source \code{habitatquarries}.
\item \code{type}: habitat type listed in \code{\link{types}} - in this
case either \code{8310} or missing (\code{NA}).
\item \code{extra_reference}: site-specific bibliographic reference(s).
Values refer to rows in the non-spatial dataframe \code{extra_references}.
}
\item \emph{if \code{references = TRUE}:} a list with both the \code{sf}
object (element \code{habitatquarries}) and a dataframe of bibliographic
references (element \code{extra_references}).
\item \emph{if \code{bibtex = TRUE}:} \code{NULL} (invisibly).
}
}
\description{
Returns the raw data source \code{habitatquarries} as an \code{sf} polygon
layer in the Belgian Lambert 72 CRS (EPSG-code
\href{https://epsg.io/31370}{31370}).
Optionally, associated bibliographic references can be returned (arguments
\code{references} or \code{bibtex}).
}
\details{
The data source \code{habitatquarries} is a GeoPackage, available at
\href{https://doi.org/10.5281/zenodo.4072967}{Zenodo}, that contains:
\itemize{
\item{\code{habitatquarries}: a spatial polygon layer that corresponds
with the presence or absence of the Natura 2000 Annex I habitat type \code{8310}
(Caves not open to the public) in the Flemish Region (and border areas),
Belgium;}
\item{\code{extra_references}: a non-spatial table of site-specific
bibliographic references.}
}

In general, different polygons represent different quarry units with their
own internal climatic environment.
Units that cross Flemish borders have been split into separate polygons.
Exceptionally they may overlap if such units are situated above each other.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'habitatquarries'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hq <- read_habitatquarries()
hq
hq2 <- read_habitatquarries(filter_hab = TRUE)
hq2
hq3 <- read_habitatquarries(references = TRUE)
hq3
read_habitatquarries(bibtex = TRUE)}

}
