% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatsprings}
\alias{read_habitatsprings}
\title{Return the data source \code{habitatsprings} as an \code{sf} point
layer}
\usage{
read_habitatsprings(
  file = file.path(locate_n2khab_data(), "10_raw/habitatsprings/habitatsprings.geojson"),
  filter_hab = FALSE,
  units_7220 = FALSE,
  version = c("habitatsprings_2025v2", "habitatsprings_2020v2")
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{filter_hab}{If \code{TRUE}, only points with (potential) habitat
are returned. The default value is \code{FALSE}.}

\item{units_7220}{If \code{TRUE}, an \code{sf} object of type-\code{7220}-locations is
returned at the population unit level.
To accomplish this, the data source is aggregated by \code{unit_id}.
Multiple points belonging to the same unit are replaced by their
centroid, their area attribute is summed (if all values are known)
and for other attributes the maximum value is returned.}

\item{version}{Version ID of the data source.
Defaults to the latest available version defined by the package.}
}
\value{
A Simple feature collection of
type \code{POINT}, with attribute variables:
\itemize{
\item \code{point_id}
\item \code{name}: site name.
\item \code{system_type}: environmental typology of \code{7220}: \code{mire},
\code{rivulet} or \code{unknown} (non-\code{7220} types are \code{NA} in older versions)
\item \code{code_orig}: original type code in raw
\code{habitatsprings}.
\item \code{type}: habitat type listed in \code{\link{types}}.
\item \code{certain}: \code{TRUE} when the type is certain and
\code{FALSE} when the type is uncertain.
\item \code{unit_id}: population unit id for large scale sampling
events.
Spatially close points have the same value.
\item \code{area_m2}: area as square meters.
\item \code{year}: year of field inventory.
\item \code{in_sac}: logical.
Is the site situated within a Special Area of Conservation?
\item \code{source}: original data source of the record.
}

Note that the \code{type} and \code{system_type} variables can have
implicit \code{NA} values
(i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can add
it as a level with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.

With \code{units_7220 = TRUE}, variable \code{point_id} is dropped and
an extra attribute variable \code{nr_of_points} is
added.
It represents the number of points that belong to each unit.
}
\description{
Returns the raw data source \code{habitatsprings} as an \code{sf} point
layer in the Belgian Lambert 72 CRS (EPSG-code
\href{https://epsg.io/31370}{31370}).
Optionally, a derived \code{sf} object of type-\code{7220}-locations can be
returned at the population unit level, through aggregation by \code{unit_id}.
}
\details{
The data source \code{habitatsprings} is a GeoJSON file (conforming to
the RFC7946 specifications), available at
\href{https://doi.org/10.5281/zenodo.3550994}{Zenodo}.
It represents sites that correspond with presence or absence
of the Natura 2000 habitat type \code{7220} (Petrifying springs with tufa
formation (\emph{Cratoneurion})) in springs and streaming water segments in
the Flemish Region, Belgium.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'habitatsprings'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hs <- read_habitatsprings()
hs
hs2 <- read_habitatsprings(units_7220 = TRUE)
hs2
}

}
