% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{read_namelist}
\alias{read_namelist}
\title{Return the \code{namelist} data source as a tibble}
\usage{
read_namelist(
  path = pkgdatasource_path("textdata/namelist", ".yml"),
  file = "namelist",
  lang = "en"
)
}
\arguments{
\item{path}{Location of the data source.
The default is to use the location of the data source as delivered by
the installed package.}

\item{file}{The filename of the data source, without extension.
The default is to use the file delivered by the installed package.}

\item{lang}{An
\href{https://www.w3.org/International/articles/language-tags/index.en}{IETF BCP
47 language tag}, such as \code{"en"} or \code{"nl"}, to specify
the language of \code{name} and \code{shortname} to be returned in the tibble.
If \code{lang = "all"}, the full \code{namelist} tibble is returned, i.e.
containing all languages.}
}
\value{
The \code{namelist} data frame as a \code{\link[tibble:tbl_df-class]{tibble}},
filtered according to the \code{lang} argument.
See \code{\link{namelist}} for documentation of the tibble's contents.
}
\description{
Returns the included data source \code{\link{namelist}} as a
\code{\link[tibble:tbl_df-class]{tibble}},
by default filtered according to English names and shortnames.
}
\details{
\code{\link{namelist}} is a data source in the
\href{https://ropensci.github.io/git2rdata}{vc-format} which provides
names and (optionally) shortnames for IDs/codes used in other data sources.

\code{read_namelist()} reads it and returns it as a
\code{\link[tibble:tbl_df-class]{tibble}}.
A tibble is a data frame that makes working in the tidyverse a little
\href{https://r4ds.had.co.nz/tibbles.html}{easier}.
By default, the data version delivered with the package is used and only English
names (\code{lang = "en"}) are returned.
}
\section{Recommended usage}{


  \code{read_namelist()}

  \code{read_namelist(lang = "nl")}
}

\examples{
read_namelist()
read_namelist(lang = "nl")

}
\seealso{
\code{\link{namelist}}

Other reading functions for n2khab-referencelists:
\code{\link[=read_env_pressures]{read_env_pressures()}},
\code{\link[=read_types]{read_types()}}
}
\concept{reading functions for n2khab-referencelists}
