% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{read_schemes}
\alias{read_schemes}
\title{Return the 'schemes' data source as a tibble with names & shortnames}
\usage{
read_schemes(path = pkgdatasource_path("textdata/schemes", ".tsv"),
  file = "schemes", file_namelist = "namelist", lang = "en")
}
\arguments{
\item{path}{Location of the data sources \code{schemes} and \code{namelist}.
The default is to use the location of the data sources as delivered by
the installed package.}

\item{file}{The filename of the \code{schemes} data source, without extension.
The default is to use the file delivered by the installed package.}

\item{file_namelist}{The filename of the \code{namelist} data source,
without extension.
The default is to use the file delivered by the installed package.}

\item{lang}{An
\href{https://www.w3.org/International/articles/language-tags/index.en}{
IETF BCP 47 language tag}, such as \code{"en"} or \code{"nl"}, to specify
the language of names & shortnames to be returned in the tibble.}
}
\value{
The \code{schemes} dataframe as a \code{\link[tibble:tbl_df-class]{tibble}},
with names & shortnames added for scheme, programme, attributes and tags
according to the \code{lang} argument.
The tibble has 25 variables.
See \code{\link{schemes}} for documentation of the data-source's contents.
See \code{\link{namelist}} for the link between codes or other identifiers
and the corresponding names (and shortnames).

The added names and shortnames are represented by the following variables:
\itemize{
  \item \code{scheme_name}
  \item \code{scheme_shortname}
  \item \code{programme_name}
  \item \code{attribute_1_name}
  \item \code{attribute_1_shortname}
  \item \code{attribute_2_name}
  \item \code{attribute_2_shortname}
  \item \code{attribute_3_name}
  \item \code{attribute_3_shortname}
  \item \code{tag_1_name}
  \item \code{tag_1_shortname}
  \item \code{tag_2_name}
  \item \code{tag_2_shortname}
  \item \code{tag_3_name}
  \item \code{tag_3_shortname}
}

The added names and shortnames for scheme, programme and attributes are
\emph{factors} with their level order according to that of the
scheme, programme or attribute variable.
}
\description{
Returns the included data source \code{\link{schemes}} as a
\code{\link[tibble:tbl_df-class]{tibble}}.
Names and shortnames from \code{\link{namelist}} are added,
in English by default.
}
\details{
\code{\link{schemes}} is a data source in the
\href{https://inbo.github.io/git2rdata/index.html}{vc-format} which provides
a list of (monitoring) schemes for N2KHAB monitoring programmes or
other N2KHAB projects, together
with defining attributes and optional information.
A 'scheme' refers to a monitoring or research setup that determines
which types (habitat/RIBs) are to be investigated for a question or for
a bunch of related questions.

\code{read_schemes()} reads the \code{\link{schemes}} data source, adds
names + shortnames and returns it as a
\code{\link[tibble:tbl_df-class]{tibble}}.
A tibble is a dataframe that makes working in the tidyverse a little
\href{https://r4ds.had.co.nz/tibbles.html}{easier}.
By default, the data version delivered with the package is used and English
names (\code{lang = "en"}) are returned for scheme, programme, attributes
and tags.
}
\section{Recommended usage}{


  \code{read_schemes()}

  \code{read_schemes(lang = "nl")}
}

\examples{
\dontrun{
read_schemes()
read_schemes(lang = "nl")
}

}
\seealso{
\code{\link{schemes}}

Other reading functions for n2khab-referencelists: \code{\link{read_env_pressures}},
  \code{\link{read_namelist}}, \code{\link{read_types}}
}
\concept{reading functions for n2khab-referencelists}
