% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilmap.R
\name{read_soilmap}
\alias{read_soilmap}
\title{Return the \code{soilmap} or \code{soilmap_simple} data source as an
\code{sf} multipolygon layer}
\usage{
read_soilmap(
  file = file.path(locate_n2khab_data(),
    "20_processed/soilmap_simple/soilmap_simple.gpkg"),
  file_raw = file.path(locate_n2khab_data(), "10_raw/soilmap"),
  use_processed = TRUE,
  version_processed = "soilmap_simple_v2",
  standardize_coastalplain = FALSE,
  simplify = FALSE,
  explan = FALSE
)
}
\arguments{
\item{file}{The absolute or relative file path of the \emph{processed} data
source \code{soilmap_simple}.
Used only if \code{use_processed = TRUE} (= default).
The default value follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file_raw}{Same as \code{file}, to define the filepath of the \emph{raw}
datasource \code{soilmap}.
Used only if \code{use_processed = FALSE}.}

\item{use_processed}{Logical.
If \code{TRUE} (the default), load and return the processed data source
\code{soilmap_simple}, instead of the raw data source \code{soilmap}.
The central layer of
\code{soilmap_simple} can be manually generated by reading the raw \code{soilmap} data
source with \code{standardize_coastalplain=TRUE}, \code{simplify=TRUE} and
\code{explan=FALSE}, but this takes some time.}

\item{version_processed}{Version ID of the \code{soilmap_simple} data source.
Only used with \code{use_processed = TRUE} (the default).
Defaults to the latest available version defined by the package.}

\item{standardize_coastalplain}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, fill the values of the morphogenetic substrate,
texture and drainage variables (\code{bsm_mo_substr}, \code{bsm_mo_tex} and
\code{bsm_mo_drain} + their \verb{_explan} counterparts) where possible,
\emph{for features with a geomorphological soil type code}.
This largely applies to features in the 'coastal plain' area.
\itemize{
\item To derive morphogenetic texture and drainage levels from the geomorphological
soil types, a conversion table by Bruno De Vos &
Carole Ampe is applied (for earlier work on this, see Ampe 2013).
\item Substrate classes are copied over from \code{bsm_ge_substr} into \code{bsm_mo_substr}
(\code{bsm_ge_substr} already follows the categories of \code{bsm_mo_substr}).
\item A logical variable is added to the output to mark conversions
(see section \emph{Value}).
}

These steps coincide with the approach that was taken to construct
\code{bsm_mo_soilunitype} in the raw data source.}

\item{simplify}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, only a limited number of variables that are most
useful for analytical work are returned.}

\item{explan}{Logical, defaults to \code{FALSE}.
Should the \verb{_explan} variables accompanying \code{bsm_mo_xxx} variables be
returned in the \emph{simplified} result?
If \code{use_processed = FALSE}: only has effect if \code{simplify=TRUE}.
(With \code{simplify=FALSE}, the \verb{_explan} variables are always returned.)
If \code{use_processed = TRUE}: is applied in returning \code{soilmap_simple}.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON},
representing either the processed data source \code{soilmap_simple} (default) or
the raw data source \code{soilmap}.

Besides the standardization for the coastal plain areas, \code{soilmap_simple}
contains
only a subset of the \code{soilmap} variables (marked with an asterisk below).

The \code{soilmap} attribute variables all start with prefix \code{bsm_} (referring
to the 'Belgian soil map'), in order to distinguish from similar attributes
derived from other maps or field observations.

Most attributes represent categories and are returned as
factors.
When a variable is a one-to-one translation of another (e.g. code vs.
explanation), the order of factor levels is aligned.

Three types of data frame variables are returned when reading \code{soilmap}:
\itemize{
\item \strong{variables with \code{mo_} in their name}: their categories follow the
Belgian Morphogenetic System.
\itemize{
\item With \code{standardize_coastalplain = FALSE}, these are only available \emph{outside
the coastal plain areas} except for \code{bsm_mo_soilunitype} (which is
standardized already in the raw data source).
}
\item \strong{variables with \code{ge_} in their name}: their categories follow the
Belgian Geomorphological System.
(Note however, that \code{bsm_ge_substr} does follow the Belgian Morphogenetic
System as well.)
\itemize{
\item Values are typically available \emph{within the coastal plain areas},
but some geomorphological soil types (starting with letter \code{O})
have a wider distribution across Flanders.
\item They are \emph{not} included in \code{soilmap_simple}.
\item A special variable is \code{bsm_ge_typology}, which is \code{TRUE} if
\code{bsm_soiltype} follows the geomorphological typology,
and \code{FALSE} otherwise.
}
\item \strong{variables without \code{mo_} or \code{ge_} in their name} are:
\itemize{
\item either \emph{system-agnostic} metadata (first two + last four variables:
\code{bsm_poly_id}, \code{bsm_map_id}, \code{bsm_map_url}, \code{bsm_book_url},
\code{bsm_detailmap_url}, \code{bsm_profloc_url}),
\item or \emph{mixed} (representing \code{mo_} categories within and \code{ge_} categories
outside coastal plains): the other ones, like \code{bsm_region}, \code{bsm_legend},
\code{bsm_soiltype} and \code{bsm_soilseries}.
\item A special variable is \code{bsm_converted}, returned only if
\code{standardize_coastalplain = TRUE}.
}
}

Many variables have a 'counterpart variable' with suffix \verb{_explan}:
they provide a more elaborate textual explanation.
They are not listed below.

Short explanation of attributes is given below.
More elaborate explanations can be found in the references and in metadata
\href{https://www.dov.vlaanderen.be/geonetwork/srv/dut/catalog.search#/metadata/5c129f2d-4498-4bc3-8860-01cb2d513f8f}{at DOV}.
\enumerate{
\item Meaning of the main non-metadata variables:
\itemize{
\item \code{bsm_region} (*): name of the region
\item \code{bsm_ge_region}: code of the region within the coastal plain area
\item \code{bsm_legend}: generalised (simplified) legend key (37 levels)
\item \code{bsm_legend_title} and \code{bsm_legend_explan}:
the legend keys and text of Van Ranst & Sys (2000) (833 and 622 levels,
respectively)
\item \code{bsm_soiltype}: the soil type of the Belgian soil map (mixed nature:
morphogenetic & geomorphological codes).
\code{bsm_soiltype_id} represents a numeric code for each level.
\item \code{bsm_ge_typology}: Logical.
Does the soiltype code follow the geomorphological typology?
\item \code{bsm_soiltype_region}: \code{bsm_soiltype}, followed by a code representing
\code{bsm_region}
\item \code{bsm_soilseries}: either the morphogenetic soil series (outside the
coastal plain areas),
which is the three core characters of \code{bsm_soiltype},
or just \code{bsm_soiltype} if the latter has a geomorphological code.
\item \code{bsm_converted} (*): Logical.
Were morphogenetic texture and drainage variables (\code{bsm_mo_tex} and
\code{bsm_mo_drain}) derived from a conversion table?
This is equivalent with the question: does \code{bsm_mo_soilunitype} differ
from \code{bsm_soiltype}?
Value \code{TRUE} is largely confined to the 'coastal plain' areas.
Only returned if \code{standardize_coastalplain = TRUE}.
(Note: the variable is not included in version \code{soilmap_simple_v1}.)
\item \code{bsm_mo_soilunitype} (*): as \code{bsm_soiltype}, but applying morphogenetic
codes within the coastal plain areas in most cases
(see the \code{standardize_coastalplain}
argument for more information about this conversion)
\item \code{bsm_mo_substr} (*), \code{bsm_ge_substr}: code of the soil substrate
\item \code{bsm_mo_tex} (*): code of the soil texture category
\item \code{bsm_mo_drain} (*): code of the soil drainage category
\item \code{bsm_mo_prof} (*): code of the soil profile category
\item \code{bsm_mo_parentmat} (*): code of a variant regarding the parent material
\item \code{bsm_mo_profvar} (*): code of a variant regarding the soil profile
\item \code{bsm_mo_phase}: code of the soil phase (i.e. additional soil
properties).
They are explained in the book that accompanies the specific analog map
identified by \code{bsm_map_id}.
\item \code{bsm_ge_series}: the geomorphological soil series
\item \code{bsm_ge_subseries}: the geomorphological soil subseries
}
\item Meaning of the metadata variables:
\itemize{
\item \code{bsm_poly_id} (*): unique polygon ID (numeric)
\item \code{bsm_map_id}: code of the analog map covering this area
\item \code{bsm_map_url}: hyperlink to the scanned analog map scale 1:20000 (pdf),
identified by \code{bsm_map_id}
\item \code{bsm_bookurl}: hyperlink to the scanned book (pdf), accompanying the
analog map identified by \code{bsm_map_id}
\item \code{bsm_detailmap_url}: hyperlink to the scanned maps at scale 1:5000
(zip-file with jpg files) belonging to the map identified by
\code{bsm_map_id}
\item \code{bsm_profloc_url}: hyperlink to the scanned maps with the profile
locations
(zip-file with jpg files) belonging to the map identified by
\code{bsm_map_id}
}
}

(*) Included in the \code{soilmap_simple} data source.
}
\description{
Returns either the raw data source \code{soilmap}
or (by default) the
processed data source \code{soilmap_simple}
as a standardized \code{sf} multipolygon layer (tidyverse-styled,
internationalized) in the Belgian Lambert 72 CRS (EPSG-code
\href{https://epsg.io/31370}{31370}).
Given the size of these data sources (especially the raw one), this function
takes a bit longer than usual to run.
}
\details{
The raw data source is published
\href{https://www.dov.vlaanderen.be/geonetwork/srv/dut/catalog.search#/metadata/5c129f2d-4498-4bc3-8860-01cb2d513f8f}{at DOV}
(Databank Ondergrond Vlaanderen) and
is discussed by Van Ranst & Sys (2000) and Dudal et al. (2005).
A 'pure' (single) dataformat of the raw data source (no metadatafiles etc.)
has also been stored (with versioning) at
Zenodo (\doi{10.5281/zenodo.3387007}) - which we refer to as the
\code{soilmap} data source - in order
to support the \code{read_soilmap()} function and to sustain long-term workflow
reproducibility.

The processed data source \code{soilmap_simple} is a GeoPackage, available at
\href{https://doi.org/10.5281/zenodo.3732903}{Zenodo}.

Note that factors are generated with implicit \code{NA} values (i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can convert
such variables with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.

In case the raw data source \code{soilmap} is used
(\code{use_processed = FALSE}), it is possible to
manually perform the standardization for coastal plain features and/or the
simplification,
both of which were applied in the \code{soilmap_simple} data source.
See \emph{Arguments} for more information.

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation of the
\code{soilmap_simple} data source from
the \code{soilmap} data source.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'soilmap_simple'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

library(dplyr)
soilmap_simple <- read_soilmap()
soilmap_simple
soilmap_simple \%>\%
  filter(!is.na(bsm_mo_substr)) \%>\%
  glimpse()
soilmap_simple \%>\%
  filter(bsm_converted) \%>\%
  glimpse()
}

}
\references{
\itemize{
\item Ampe C. (2013). Databank aardewerk Vlaanderen 2010.
Omzetten (zeer) oude legende bodemkartering naar legende bodemkaart
Kuststreek. Vlaamse Landmaatschappij Regio West, Bruges, 45 p.
\item Dudal R., Deckers J., Van Orshoven J. & Van Ranst E. (2005). Soil survey
in Belgium and its applications. In: Bullock P., Jones R.J.A., Montanarella
L. (editors). Soil Resources of Europe. Office for Official Publications of
the European Communities, Luxembourg, p. 63–71.
URL: \url{http://hdl.handle.net/1854/LU-368514}.
\item Van Ranst E. & Sys C. (2000). Eenduidige legende van de digitale
bodemkaart van Vlaanderen (schaal 1: 20000). Universiteit Gent, Laboratorium
voor Bodemkunde, Ghent, 361 p. URL: \url{http://hdl.handle.net/1854/LU-125899}.
}
}
\seealso{
Other functions returning environmental data sets:
\code{\link[=read_shallowgroundwater]{read_shallowgroundwater()}},
\code{\link[=read_watercourse_100mseg]{read_watercourse_100mseg()}},
\code{\link[=read_watersurfaces]{read_watersurfaces()}}
}
\concept{functions returning environmental data sets}
