% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{read_types}
\alias{read_types}
\title{Return the \code{types} data source as a tibble with names & shortnames}
\usage{
read_types(
  path = pkgdatasource_path("textdata/types", ".yml"),
  file = "types",
  file_namelist = "namelist",
  lang = "en"
)
}
\arguments{
\item{path}{Location of the data sources \code{types} and \code{namelist}.
The default is to use the location of the data sources as delivered by
the installed package.}

\item{file}{The filename of the \code{types} data source, without extension.
The default is to use the file delivered by the installed package.}

\item{file_namelist}{The filename of the \code{namelist} data source,
without extension.
The default is to use the file delivered by the installed package.}

\item{lang}{An
\href{https://www.w3.org/International/articles/language-tags/index.en}{
IETF BCP 47 language tag}, such as \code{"nl"} or \code{"en"}, to specify
the language in which the columns of the tibble should be returned.}
}
\value{
The \code{types} data frame as a \code{\link[tibble:tbl_df-class]{tibble}},
with names & shortnames added for types, attributes and tags
according to the \code{lang} argument.
See \code{\link{types}} for documentation of the data-source's contents.
See \code{\link{namelist}} for the link between codes or other identifiers
and the corresponding names (and shortnames).

The added names and shortnames are represented by the following variables:
\itemize{
  \item \code{type_name}
  \item \code{type_shortname}
  \item \code{typeclass_name}
  \item \code{hydr_class_name}
  \item \code{hydr_class_shortname}
  \item \code{groundw_dep_name}
  \item \code{groundw_dep_shortname}
  \item \code{flood_dep_name}
  \item \code{flood_dep_shortname}
  \item \code{tag_1_name}
  \item \code{tag_1_shortname}
  \item \code{tag_2_name}
  \item \code{tag_2_shortname}
  \item \code{tag_3_name}
  \item \code{tag_3_shortname}
}
Except for the tags, the names and shortnames are factors with their level
order according to that of the corresponding attribute.
}
\description{
Returns the included data source \code{\link{types}} as a
\code{\link[tibble:tbl_df-class]{tibble}}.
Names and shortnames from \code{\link{namelist}} are added,
in English by default.
}
\details{
\code{\link{types}} is a data source in the
\href{https://ropensci.github.io/git2rdata}{vc-format} which provides
a checklist of types, represented by their \strong{current} codes, together
with several attributes.
A 'type' refers to either a (main) habitat type, a
habitat subtype or a regionally important biotope (RIB).

\code{read_types()} reads the \code{\link{types}} data source, adds
names + shortnames
and returns it as a
\code{\link[tibble:tbl_df-class]{tibble}}.
A tibble is a data frame that makes working in the tidyverse a little
\href{https://r4ds.had.co.nz/tibbles.html}{easier}.
By default, the data version delivered with the package is used and English
names (\code{lang = "en"}) are returned for types, attributes and tags.

Note that factors are generated with implicit \code{NA} values (i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can convert
such variables with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.
}
\section{Recommended usage}{


  \code{read_types()}

  \code{read_types(lang = "nl")}
}

\examples{
read_types()
read_types(lang = "nl")

}
\seealso{
\code{\link{types}}

Other reading functions for n2khab-referencelists:
\code{\link[=read_env_pressures]{read_env_pressures()}},
\code{\link[=read_namelist]{read_namelist()}}
}
\concept{reading functions for n2khab-referencelists}
