% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata.R
\name{scheme_types}
\alias{scheme_types}
\title{Documentation of included data source 'scheme_types'}
\format{
A vc-formatted data source. As such, it corresponds to
a dataframe with 3 variables:
\describe{
  \item{scheme}{Code of the scheme, as a factor, and explained by
  \code{\link{namelist}}.
  It must be represented in \code{\link{schemes}}.
  }
  \item{type}{Code of the type, as a factor, and explained by
  \code{\link{namelist}}.
  It must be represented in \code{\link{types}}.}
  \item{typegroup}{An \emph{optional} code (and optionally explained by
  \code{\link{namelist}}),
  declaring the typegroup to which a type belongs \emph{within the specified
  scheme}.
  Typegroups point out which types are considered together as a group in
  inferences for the scheme at hand.}
}

Each combination of \code{scheme} and \code{type} must be unique.
}
\source{
\itemize{
    \item{For the MNE-schemes:
      \itemize{
        \item{the link between schemes and types comes from
        a vc-formatted, Dutch-language data source
        which can be found in the
        \href{https://github.com/inbo/n2khab}{
        n2khab Github repository}:
        \code{misc/generate_textdata/rawraw_data/10_compmeetnet_types_milieudrukken}.
        This data source was originally generated in the
        n2khab-mne-selections repository
        (\samp{https://gitlab.com/florisvdh/n2khab-mne-selections}).}
        \item{the declaration of typegroups comes from
        \href{https://docs.google.com/spreadsheets/d/1n2ohvuLEK_anX37gxlanQXRGotweMVonFxKxtfTcuZI}{
        this googlesheet}.
        Currently, the typegroups are kept up to date both in the
        googlesheet and in the data source.}
        }
    }
  }

\itemize{
    \item{For the MHQ-schemes, the list of sampled types is based on the report of
    \href{https://pureportal.inbo.be/portal/files/4339795/Westra_etal_2014_MonitoringNatura2000Habitats.pdf}{Westra \emph{et al.} (2014)}.
  }
 }
}
\description{
'scheme_types' is a data source in the
\href{https://ropensci.github.io/git2rdata}{vc-format} which lists
the types (using the type-code from \code{\link{types}}) that belong to
each N2KHAB (monitoring or research) scheme (using the scheme-code from
\code{\link{schemes}}).
It also defines typegroup memberships of the types within specific schemes,
if applicable.
The codes of schemes and types (and optionally: typegroups)
are explained in the
data source \code{\link{namelist}} (which can accommodate multiple
languages).
}
\section{Typical way of loading}{


  \code{read_scheme_types()}

  \code{read_scheme_types(lang = "nl")}
}

\section{Corresponding datafiles in the installed package}{


  \code{textdata/scheme_types.tsv}

  \code{textdata/scheme_types.yml}
}

\seealso{
\code{\link{read_scheme_types}}

Other n2khab-referencelists: 
\code{\link{env_pressures}},
\code{\link{namelist}},
\code{\link{schemes}},
\code{\link{types}}
}
\concept{n2khab-referencelists}
