% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata.R
\name{schemes}
\alias{schemes}
\title{Documentation of included data source 'schemes'}
\format{
A vc-formatted data source. As such, it corresponds to
a dataframe with 10 variables:
\describe{
  \item{scheme}{Code of the scheme, as a factor.
  This is the ID for use in diverse workflows and datasets.
  Corresponding names and shortnames
  in multiple languages are to be found in \code{\link{namelist}}.
  Contains no duplicates!

  A 'scheme' refers to a monitoring or research setup that determines
  which types (habitat/RIBs) are to be investigated for a question or for
  a bunch of related questions.
  }
  \item{programme}{Code of the programme to which the scheme belongs.
  The programme's code is explained by \code{\link{namelist}}.
  Is a factor.

  A 'programme' refers to a N2KHAB monitoring programme or research project.
  One programme can correspond to multiple schemes.
  At least the monitoring programmes MHQ and MNE are present.
  }
  \item{attribute_1}{The first defining attribute of the scheme.
  Typically, the code is explained by \code{\link{namelist}}.
  Is a factor.
  \itemize{
    \item{In MNE, this is used to declare the 'compartment superscheme' (scheme
    collection) to which the scheme belongs, and which is named after the
    environmental compartment (however note that the surfacewater superscheme
    comprises two environmental compartments).}
    \item{In MHQ, this is used to define the target habitat type of the
    monitoring scheme.}
  }
  }
  \item{attribute_2}{A second defining attribute of the scheme (if needed).
  Typically, the code is explained by \code{\link{namelist}}.
  Is a factor.
  \itemize{
    \item{In MNE, this provides the code of the environmental pressure to
    which the scheme is related.
    It must be represented in \code{\link{env_pressures}}.}
  }
  }
  \item{attribute_3}{A third defining attribute of the scheme (if needed).
  Typically, the code is explained by \code{\link{namelist}}.
  Is a factor.
  \itemize{
    \item{In MNE, this is used to declare an \strong{optional} partitioning of the
    scheme that would be constituted by attributes 1 and 2 alone
    (environmental compartment and pressure), which will hence divide the
    concerned types in at least two groups.
    This typically has to do with differences in the considered
    environmental variables or compatibility of interpretation.

    It is \emph{not} meant to distinguish between different types or
    typegroups for inferences, so types are kept in the same scheme as
    long as the measured values have compatibility of interpretation.
    To declare which types or typegroups should be
    distinguished in inferences, use the \code{typegroup} variable in the
    \code{\link{scheme_types}} data source!}
  }
  }
  \item{spatial_restriction}{Optional further defining specification of a
  scheme (in English):
  spatial restrictions, superposed on the restrictions already
  generated by the attributes.}
  \item{notes}{Optional. In English. For additional notes on the scheme
  definition, if necessary.}
  \item{tag_1}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.
  \itemize{
    \item{In MNE, this is used to tag which schemes are focal schemes and
    which ones are secondary schemes.}
    \item{In MHQ, this is used to separate aquatic and terrestial
    monitoring schemes.}
  }
  }
  \item{tag_2}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.}
  \item{tag_3}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.}
}
}
\source{
\itemize{
    \item{For the MNE-schemes: a vc-formatted, Dutch-language data source
    which can be found in the
    \href{https://github.com/inbo/n2khab}{
    n2khab Github repository}:
    \code{misc/generate_textdata/rawraw_data/10_compmeetnet_types_milieudrukken}.
    This data source was originally generated in the
    n2khab-mne-selections repository
    (\samp{https://gitlab.com/florisvdh/n2khab-mne-selections}).
    }
  }
\itemize{
    \item{More information on the MHQ-schemes can be found in
    \href{https://pureportal.inbo.be/portal/files/4339795/Westra_etal_2014_MonitoringNatura2000Habitats.pdf}{Westra \emph{et al.} (2014)}.
  }
 }
}
\description{
'schemes' is a data source in the
\href{https://ropensci.github.io/git2rdata}{vc-format} which provides
a list of (monitoring) schemes for N2KHAB monitoring programmes or
other N2KHAB projects, together
with defining attributes and optional information.
The codes of schemes, programmes, attributes and tags
are explained in the
data source \code{\link{namelist}} (which can accommodate multiple
languages).
}
\section{Typical way of loading}{


  \code{read_schemes()}

  \code{read_schemes(lang = "nl")}
}

\section{Corresponding datafiles in the installed package}{


  \code{textdata/schemes.tsv}

  \code{textdata/schemes.yml}
}

\seealso{
\code{\link{read_schemes}}

Other n2khab-referencelists: 
\code{\link{env_pressures}},
\code{\link{namelist}},
\code{\link{scheme_types}},
\code{\link{types}}
}
\concept{n2khab-referencelists}
