% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{url_is_up}
\alias{url_is_up}
\title{Check availability of a URL}
\usage{
url_is_up(url, timeout = 2)
}
\value{
A logical of length 1.
}
\description{
Check availability of a URL
}
\details{
The code is based on \url{https://stackoverflow.com/a/60627969}.
}
\keyword{internal}
