% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{checksum}
\alias{checksum}
\alias{md5sum}
\alias{sha256sum}
\title{Calculate file checksums (OpenSSL implementation)}
\usage{
md5sum(files)

sha256sum(files)
}
\arguments{
\item{files}{Character vector of file path(s).
File path(s) can be absolute or relative.}
}
\value{
Character vector with the same length as \code{file}.
}
\description{
The functions calculate the checksum (cryptographic digest, hash value) of
one or multiple files.
They can be used to verify file integrity.
}
\details{
The function names were chosen to match those of GNU coreutils.

The functions use the OpenSSL implementation and stream-hash the binary
contents of the connections to the respective files.
They turn the hash-format for binary streams by the \code{openssl} package
into a regular hash string.
Note that \code{n2khab} will mask
\code{\link[tools:md5sum]{tools::md5sum()}},
which is a standalone implementation.
}
\examples{
# creating two different temporary files:
file1 <- tempfile()
file2 <- tempfile()
files <- c(file1, file2)
file.create(files)
con <- file(file2)
writeLines("some text", con)
close(con)

# computing two alternative checksums:
md5sum(files)
sha256sum(files)

\dontrun{
# This will error:
files <- c(file1, file2, tempfile(), tempfile())
md5sum(files)
}

}
\seealso{
Other functions regarding file management for N2KHAB projects: 
\code{\link{download_zenodo}()},
\code{\link{fileman_folders}()},
\code{\link{fileman_up}()}
}
\concept{functions regarding file management for N2KHAB projects}
