% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{checksum}
\alias{checksum}
\alias{xxh64sum}
\alias{md5sum}
\alias{sha256sum}
\title{Calculate file checksums}
\usage{
checksum(files, hash_fun = c("xxh64", "md5", "sha256"))

xxh64sum(files)

md5sum(files)

sha256sum(files)
}
\arguments{
\item{files}{Character vector of file path(s).
File path(s) can be absolute or relative.}

\item{hash_fun}{String that defines the hash function.
See \emph{Usage} for allowed values; defaults to the first.}
}
\value{
Named character vector with the same length as \code{files}
and with the file names as names.
}
\description{
The functions calculate the checksum (digest; hash value) of
one or multiple files.
They can be used to verify file integrity.
}
\details{
A few cryptographic and non-cryptographic hash functions are implemented,
either from the OpenSSL library (through
\href{https://CRAN.R-project.org/package=openssl}{\code{openssl}})
or as embedded in the
\href{https://CRAN.R-project.org/package=digest}{\code{digest}}
package.

Functions \code{md5sum()} etc. are simple shortcuts to \code{checksum()}
with the appropriate hash function preset.
Their names were chosen to match those of xxHash and GNU coreutils.

The cryptographic algorithms use the OpenSSL implementation and
stream-hash the binary
contents of the connections to the respective files.
They turn the hash-format for binary streams by the \code{openssl} package
into a regular hash string.
Note that \code{n2khab} will mask
\code{\link[tools:md5sum]{tools::md5sum()}},
which is a standalone implementation.
}
\examples{
# creating two different temporary files:
file1 <- tempfile()
file2 <- tempfile()
files <- c(file1, file2)
file.create(files)
con <- file(file2)
writeLines("some text", con)
close(con)

# computing alternative checksums:
checksum(files)
xxh64sum(files)
md5sum(files)
sha256sum(files)

\dontrun{
# This will error:
files <- c(file1, file2, tempfile(), tempfile())
checksum(files)
}

}
\seealso{
Other functions regarding file management for N2KHAB projects: 
\code{\link{download_zenodo}()},
\code{\link{fileman_folders}()},
\code{\link{fileman_up}()}
}
\concept{functions regarding file management for N2KHAB projects}
