% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTSmh.R
\name{convert_dec_to_base4frac}
\alias{convert_dec_to_base4frac}
\title{Convert a vector of values from the \code{GRTSmaster_habitats} data source to
base 4 fractions}
\usage{
convert_dec_to_base4frac(x)
}
\arguments{
\item{x}{A decimal (i.e. base 10) scalar or vector of integer values from the
\code{GRTSmaster_habitats} raster.}
}
\value{
The corresponding base4 scalar or vector, stored as a fraction.
}
\description{
Converts decimal (i.e. base 10) integer values from the raster data source
\code{GRTSmaster_habitats} into base 4 fractions, using a precision of
13 digits behind the decimal mark (as needed to cope with the range of
values).
For example, the integer \code{16} (\code{= 4^2}) is converted into
\code{0.0000000000100} and \code{4^12} is converted into
\code{0.1000000000000}.
}
\details{
Long base 4 fractions seem to be handled and stored easier than long
(base 4) integers.
This approach follows the one of Stevens & Olsen (2004) to represent
the reverse hierarchical order in a GRTS sample as base-4-fraction addresses.

The function works on a vector and retains \code{NA} values.
As such, it can be used in \code{raster::calc()}.
When writing such a raster to a file, it is needed to use the \code{FLT8S}
data type (see \code{\link[raster]{dataType}}).
Otherwise several digits will change.

The function is based on code from the \code{baseConvert()} function
in Will Gray's \href{https://github.com/graywh/r-gmisc}{Gmisc} package.
}
\examples{
\dontrun{
oldoption <- getOption("digits")
options(digits = 15)
convert_dec_to_base4frac(c(14, 15, NA, 456))
options(digits = oldoption)
}

}
\references{
Stevens D.L. & Olsen A.R. (2004). Spatially Balanced Sampling of Natural
Resources. Journal of the American Statistical Association 99 (465):
262–278. DOI: 10.1198/016214504000000250.
}
\seealso{
Other functions involved in processing the \code{GRTSmaster_habitats} data source: \code{\link{convert_base4frac_to_dec}},
  \code{\link{read_GRTSmh_base4frac}},
  \code{\link{read_GRTSmh_diffres}},
  \code{\link{read_GRTSmh}}
}
\concept{functions involved in processing the \code{GRTSmaster_habitats} data source}
