% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{convert_line_endings}
\alias{convert_line_endings}
\title{Convert line endings of text files}
\usage{
convert_line_endings(
  files,
  direction = c("to_lf", "to_crlf"),
  root = NULL,
  force = FALSE,
  silent = TRUE
)
}
\arguments{
\item{files}{Character vector of file paths; these may be relative to \code{root}.}

\item{direction}{Whether conversion is to LF (\code{"to_lf"}) or to CLRF
(\code{"to_crlf"}).}

\item{root}{An optional directory path; if present the \code{files} argument is
considered relative to \code{root}.}

\item{force}{Logical; \code{TRUE} will only have effect in non-Windows
systems. See Details.}

\item{silent}{Logical. Whether to print the result to the console.}
}
\value{
The \code{files} argument; invisibly unless \code{silent = FALSE}.
}
\description{
Converts line endings of text files from CRLF (\verb{\\r\\n}) to LF (\verb{\\n}) or the
reverse.
}
\details{
The conversion to LF is especially helpful to maintain file integrity across
platforms in combination with a distributed version control system like
\code{git}.

To prevent unneeded rewriting, the \code{"to_lf"} direction is only executed in
Windows systems, unless \code{force = TRUE}.

To prevent \code{git} from automatically replacing line endings in specific files
when checking out new versions of those files, set LF as the required line
ending of specific file types in a \code{.gitattributes} file in the root of the
git repository, and commit this file so that collaborators use the same
setting. An example line in \code{.gitattributes} looks like this (this example
makes line endings of yml files always use LF):

\verb{*.yml text eol=lf}

Note that, with this setting, to have existing files updated by \code{git} they
first need to be removed from the working directory, after which you can do a
checkout again for these files.

The function is designed so that it can take the output of
\code{\link[git2rdata:write_vc]{git2rdata::write_vc()}} as input in a pipeline; still its \code{root} argument may
need to be repeated in such case.
}
\note{
The function borrows from \code{TAF::dos2unix()} and \code{TAF::unix2dos()},
which work on single files.
}
\examples{
\dontrun{
files <- c(file1, file2)
convert_line_endings(files)
}


}
\seealso{
Other functions regarding file management for N2KHAB projects: 
\code{\link{checksum}()},
\code{\link{download_zenodo}()},
\code{\link{fileman_folders}()},
\code{\link{fileman_up}()},
\code{\link{get_zenodo_versions}()},
\code{\link{locate_n2khab_data}()}
}
\concept{functions regarding file management for N2KHAB projects}
