% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\docType{import}
\name{download_zenodo}
\alias{download_zenodo}
\title{Get data from a Zenodo archive}
\keyword{documentation}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{inborutils}{\code{\link[inborutils]{download_zenodo}}}
}}

