% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata.R
\name{env_pressures}
\alias{env_pressures}
\title{Documentation of included data source 'env_pressures'}
\format{
A vc-formatted data source. As such, it corresponds to
a dataframe with 35 rows and 3 variables:
\describe{
  \item{ep_code}{Code of the environmental pressure, as a factor.
  This is the ID for use in diverse workflows and datasets.
  Corresponding names and abbreviations
  in multiple languages are stored in \code{\link{namelist}}
  (as name and shortname, respectively).
  The abbreviation may be seen as an alternative, language-dependent code.
  Contains no duplicates!}
  \item{ep_class}{A code explained by \code{\link{namelist}},
  corresponding to the environmental pressure's class.
  Is a factor.}
  \item{explanation}{A code explained by \code{\link{namelist}},
  corresponding to the explanation on the environmental pressure, and
  optional remarks.
  Explanation and remarks are stored in \code{\link{namelist}}
  (as name and shortname, respectively)}
}
}
\source{
The latest worksheet of
\href{https://docs.google.com/spreadsheets/d/1PH6InqJk0ijQF_N7v7IZjarijlqHRBCKhTbDn44skZU}{
this googlesheet}.
Currently, the googlesheet and the data source are both kept up-to-date.
}
\description{
'env_pressures' is a data source in the
\href{https://ropensci.github.io/git2rdata}{vc-format} which provides
a checklist of environmental pressures, represented by codes, together
with the pressure-class and the textual explanation (with optional remarks).
The codes of environmental pressures, pressure-classes and explanations
are explained in the
data source \code{\link{namelist}} (which can accommodate multiple
languages).
}
\section{Typical way of loading}{


  \code{read_env_pressures()}

  \code{read_env_pressures(lang = "nl")}
}

\section{Corresponding datafiles in the installed package}{


  \code{textdata/env_pressures.tsv}

  \code{textdata/env_pressures.yml}
}

\seealso{
\code{\link{read_env_pressures}}

Other n2khab-referencelists: 
\code{\link{namelist}},
\code{\link{scheme_types}},
\code{\link{schemes}},
\code{\link{types}}
}
\concept{n2khab-referencelists}
