% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{filemanag_zenodo}
\alias{filemanag_zenodo}
\title{Get raw data from a zenodo archive

This function will download data from Zenodo (\href{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)}
\usage{
filemanag_zenodo(path, doi)
}
\arguments{
\item{path}{local path to where the data need to be written}

\item{doi}{a doi pointer to the Zenodo archive starting with '10.5281/zenodo.'. See examples.}
}
\value{
Downloaded file(s) in the specified folder.
}
\description{
Get raw data from a zenodo archive

This function will download data from Zenodo (\href{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)
}
\examples{
\dontrun{
# Single zip file deposition
filemanag_zenodo(".", "10.5281/zenodo.1283345")
# Multiple files deposition
filemanag_zenodo(".", "10.5281/zenodo.1172801")
# Single pdf file deposition
filemanag_zenodo(".", "10.5281/zenodo.168478")
}
}
