% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filemanagement.R
\name{get_zenodo_versions}
\alias{get_zenodo_versions}
\title{Get version IDS from a Zenodo archive}
\usage{
get_zenodo_versions(doi)
}
\arguments{
\item{doi}{A DOI pointer to the Cite-all-versions DOI for a Zenodo archive.
A Zenodo archive starts with '10.5281/zenodo.'.}
}
\value{
A named vector. The names correspond to version IDs, the values
correspond to DOIs.
}
\description{
The function queries the Zenodo API to find the version IDs and
corresponding DOIs.
}
\seealso{
Other functions regarding file management for N2KHAB projects: 
\code{\link{checksum}()},
\code{\link{convert_line_endings}()},
\code{\link{download_zenodo}()},
\code{\link{fileman_folders}()},
\code{\link{fileman_up}()},
\code{\link{locate_n2khab_data}()}
}
\concept{functions regarding file management for N2KHAB projects}
