% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2khab-deprecated.R
\name{n2khab-deprecated}
\alias{n2khab-deprecated}
\alias{read_schemes}
\alias{read_scheme_types}
\title{Deprecated functions}
\usage{
read_schemes(...)

read_scheme_types(...)
}
\arguments{
\item{...}{arguments passed to the new function}
}
\description{
\itemize{
\item \code{\link[=read_schemes]{read_schemes()}} has moved to package \code{n2khabmon}.
Use \code{\link[n2khabmon:read_schemes]{n2khabmon::read_schemes()}} instead.
\item \code{\link[=read_scheme_types]{read_scheme_types()}} has moved to package \code{n2khabmon}.
Use \code{\link[n2khabmon:read_scheme_types]{n2khabmon::read_scheme_types()}} instead.
}
}
\keyword{internal}
