% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2khab.R
\docType{package}
\name{n2khab-package}
\alias{n2khab}
\alias{n2khab-package}
\title{n2khab: Data preparation for Flemish Natura 2000 habitat analyses}
\description{
The n2khab package is an R package with
    preprocessing functions and standard reference data, useful for 
    analyses regarding Flemish Natura 2000 habitats and regionally 
    important biotopes (RIBs).
}
\details{
See the \href{00Index.html}{Index} for vignettes and
for documentation on functions and datasets delivered with the package.

For contributing, see the README on
\href{https://github.com/inbo/n2khab/blob/master/README.md}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/n2khab}
  \item Report bugs at \url{https://github.com/inbo/n2khab/issues}
}

}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (0000-0002-6378-6229)

Authors:
\itemize{
  \item Toon Westra \email{toon.westra@inbo.be} (0000-0003-2478-9459)
  \item Hans Van Calster \email{hans.vancalster@inbo.be} (0000-0001-8595-8426)
}

Other contributors:
\itemize{
  \item Dries Adriaens \email{dries.adriaens@inbo.be} (0000-0002-7216-3111) [contributor]
  \item Cécile Herr \email{cecile.herr@inbo.be} (0000-0002-2833-6046) [contributor]
  \item Maud Raman \email{maud.raman@inbo.be} (0000-0003-2116-4963) [contributor]
  \item Jan Wouters \email{jan.wouters@inbo.be} (0000-0001-8200-9102) [contributor]
  \item Steven De Saeger \email{steven.desaeger@inbo.be} (0000-0002-5477-3063) [contributor]
  \item Nathalie Cools \email{nathalie.cools@inbo.be} (0000-0002-7059-2318) [contributor]
  \item Suzanna Lettens \email{suzanna.lettens@inbo.be} (0000-0001-5032-495X) [contributor]
  \item Arne Verstraeten \email{arne.verstraeten@inbo.be} (0000-0002-8813-1261) [contributor]
  \item Research Institute for Nature and Forest \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
