% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata.R
\name{namelist}
\alias{namelist}
\title{Documentation of datafile 'namelist'}
\format{A vc-formatted datafile with many rows and 4 variables:
\describe{
  \item{code}{A code used elsewhere.}
  \item{lang}{An
  \href{https://www.w3.org/International/articles/language-tags/index.en}{IETF BCP
  47 language tag}, such as \code{"en"} or \code{"nl"}, to identify the language of the
  corresponding value of variables \code{name} and \code{shortname}.}
  \item{name}{The name corresponding to \code{code} and \code{lang}.}
  \item{shortname}{Optionally, a shorter variant of \code{name}.} }}
\source{
\code{inst/textdata/namelist.tsv}

  \code{inst/textdata/namelist.yml}
}
\description{
A datafile in the
\href{https://inbo.github.io/git2rdata/index.html}{vc-format} which provides
names and (optionally) shortnames for IDs/codes used in other data sources.
Multiple languages are supported.
}
\section{Typical way of loading}{


  \code{read_namelist()}

  \code{read_namelist(lang = "nl")}
}

\seealso{
\code{\link{read_namelist}}
}
\concept{n2khab-referencelists}
