% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_admin_areas.R
\name{read_admin_areas}
\alias{read_admin_areas}
\title{Return one of the available geospatial data sources representing
administrative areas}
\usage{
read_admin_areas(
  file = file.path(fileman_up("n2khab_data"), c("10_raw/flanders", "10_raw/provinces",
    "10_raw/sac")),
  dsn = c("flanders", "provinces", "sac")
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{dsn}{A string, conforming to one of the data source names listed under
\emph{Usage}.
Considering the default values of the \code{path} and \code{file} arguments,
it is recommended to just use this argument.
Defaults to \code{"flanders"}.

Different data sources are handled differently by the function.
The \code{dsn} argument states which data source is requested.
It is also used to determine \code{file} if that argument is not specified,
in order to select the right data file(s) from the \code{n2khab_data} folder.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON} or
\code{POLYGON}.
}
\description{
Returns an administrative geospatial data source.
The coordinate reference system is 'BD72 / Belgian Lambert 72'
(EPSG-code \href{https://epsg.io/31370}{31370}).
}
\details{
See section \emph{Usage} to see which N2KHAB data sources are
available.
You get a list of the data source names and related information with
\code{XXXXXXXXX}.
You are referred to the raw N2KHAB-data collection
\href{https://zenodo.org/communities/n2khab-data-raw}{at Zenodo} to learn
more about these data sources.
}
\examples{
\dontrun{
flanders <- read_admin_areas(dsn = "flanders")
provinces <- read_admin_areas(dsn = "provinces")
sac <- read_admin_areas(dsn = "sac")
}

}
