% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_favenv.R
\name{read_favenv}
\alias{read_favenv}
\title{Read favourable environmental ranges for vegetation types of Flanders}
\usage{
read_favenv(
  file = file.path(locate_n2khab_data(), "10_raw/favenv",
    "VanCalster_etal_2020_Gunstig_abiotisch_bereik_per_milieuvariabele_en_habitatsubtype_v01_00.txt"),
  version = c("favenv_v1.0"),
  lang = c("nl")
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{version}{Version ID of the data source.
Defaults to the latest available version defined by the package.}

\item{lang}{An
\href{https://www.w3.org/International/articles/language-tags/index.en}{
IETF BCP 47 language tag}, such as \code{"en"} or \code{"nl"}, to specify
the language of names & shortnames to be returned in the tibble.}
}
\value{
A `tibble`.

The table contains 950 rows and 19 columns.
The first row is a column header and represents the variable name.
The following list provides a brief description for each variable:

`type`: Natura 2000 (sub)type habitat code

`subtype`: code/name of possible vegetation variants.
For aquatic types, the Flemish water type is listed
(according to Water Framework Directive-Integrated Water Policy Objectives)

`milieucompartiment`: soil, groundwater, inundation water (flooding with
water of external origin), water column/surface water, air

`variabele`: name of the environmental variable.

`afkorting`: abbreviation used for the environmental variable

`eenheid`: unit used for the environmental variable

`toetswijze_bepaling`: for aquatic types, measurements are first summarized
 using certain statistics such as percentiles, mean, etc. (in accordance with
 Water Framework Directive-Integrated Water Policy Objectives)

`teken`: range type
- LL-UL: lower limit-upper limit
- 10-90 perc: 10-90 percentile values
- min-max: minimum - maximum
- < - >: less than - greater than

`abiotisch_bereik`: overall measurement range of an environmental variable
 within which a habitat type can function sustainably:
- `abiotisch_bereik`: range as text field
- `waarde_1` - `waarde_2`: lower and upper limit as text field
- `waarde_num_1` - `waarde_num_2`: lower and upper limit as numerical value
 (where conversion was possible)

`n_gunstig`: number of test plots in favorable conservation status used for
the calculation. The definition of favorable conservation status is based on
Oosterlynck et al. (2020, version 3).

`status`: source and method by which the range was derived:
- Dp: ranges calculated with INBO data, prediction intervals, local
  conservation status (LCS) known
- Dt: ranges determined using Titan method (Baker & King 2010), LCS unknown
- Dk: ranges calculated with INBO data, quantiles, insufficient data to
  calculate prediction intervals
- Lg: ranges derived from literature (LCS clearly indicated)
- Lk: ranges derived from literature (LCS derived, based on site
  characteristics)

`referentie`: references based on which the abiotic range was determined

`opmerking`: remarks

`wijziging`: whether this record has been modified compared to Van Calster
et al. (2020) and how
}
\description{
The function reads the data into the current R session if `file` exists.
If the `file` does not exist, it will attempt to download it and read it
after the download has completed.
The original source for the data is Van Calster et al. (2020).
}
\examples{
\dontrun{
favenv <- read_favenv()
}

}
\references{
- Van Calster H., Cools N., De Keersmaeker L., Denys L., Herr C., Leyssen A.,
Provoost S., Vanderhaeghe F., Vandevoorde B., Wouters J. en M. Raman (2020).
Gunstige abiotische bereiken voor vegetatietypes in Vlaanderen. Rapporten van
het Instituut voor Natuur- en Bosonderzoek 2020 (44). Instituut voor Natuur-
en Bosonderzoek, Brussel. DOI: doi.org/10.21436/inbor.19362510D
- Oosterlynck P., De Saeger S., Leyssen A., Provoost S., Thomaes A.,
Vandevoorde B., Wouters J., & Paelinckx D. (2020). Criteria voor de
beoordeling van de lokale staat van instandhouding van de Natura2000
habitattypen in Vlaanderen. Rapporten van het Instituut voor Natuur- en
Bosonderzoek 2020 (27). Instituut voor Natuur- en Bosonderzoek, Brussel.
DOI: doi.org/10.21436/inbor.14061248
}
