% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTSmh.R
\name{read_GRTSmh}
\alias{read_GRTSmh}
\title{Return the \code{GRTSmaster_habitats} data source as a RasterLayer or a
10-layered variant as a RasterBrick}
\usage{
read_GRTSmh(path,
  file = c("10_raw/GRTSmaster_habitats/GRTSmaster_habitats.tif",
  "20_processed/GRTSmh_brick/GRTSmh_brick.tif"), brick = FALSE)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, use this argument
in scripts to set the location of the folder '\strong{\code{data}}'.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{https://github.com/inbo/n2khab-inputs}{n2khab-inputs} repository.}

\item{brick}{Logical; determines whether the RasterLayer or RasterBrick data
source is returned. See the Details section.}
}
\value{
Either a RasterLayer or a 10-layered RasterBrick, always with 21041043 cells.
}
\description{
By default, the \code{GRTSmaster_habitats} data source is returned as a
RasterLayer with decimal integer ranking numbers as values.
If \code{brick = TRUE}, a 10-layered RasterBrick is
returned (data source \code{GRTSmh_brick}; resolution 32 m)
with the decimal integer ranking numbers of 10 hierarchical levels of the
GRTS cell addresses, including the one from \code{GRTSmaster_habitats}
(with GRTS cell addresses at the resolution level).
}
\details{
The data source \code{GRTSmaster_habitats}, provided and documented in
\href{https://zenodo.org/record/2611234}{Zenodo}, is a monolayered GeoTIFF
file covering the whole of Flanders and the Brussels Capital Region at a
resolution of 32 m.
Its values are unique decimal integer ranking numbers from the GRTS
algorithm applied to the Flemish and Brussels area.
Beware that not all GRTS ranking numbers are present in the data source, as
the original GRTS raster has been clipped with the Flemish outer borders
(i.e., not excluding the Brussels Capital Region).

The GRTS algorithm uses a quadrant-recursive, hierarchically randomized
function that maps the unit square to the unit interval, resulting in a
base-4 GRTS address for each location (see
\code{\link{read_GRTSmh_base4frac}}).
The ranking numbers in \code{GRTSmaster_habitats} are base-10 numbers and
follow the reverse
hierarchical order: each consecutive subset of ranking numbers corresponds
to a spatially balanced sample of locations.
Hence, it allows dynamical sample sizes.
More information on the GRTS algorithm can be found in Stevens & Olsen (2003,
2004) and in the \href{https://github.com/ThierryO/grts}{GRTS} and
\href{https://cran.r-project.org/web/packages/spsurvey}{spsurvey} packages.

Depending on the value of the \code{brick} argument, the function either
returns the \code{GRTSmaster_habitats} data source as a
RasterLayer (\code{brick = FALSE}), or (\code{brick = TRUE}) returns the
10-layered RasterBrick \code{GRTSmh_brick} (resolution 32 m)
with the decimal integer ranking numbers of 10 hierarchical levels of the
GRTS cell addresses, including the one from \code{GRTSmaster_habitats}
(with GRTS cell addresses at the resolution level).
The \code{GRTSmh_brick} data source is a processed dataset (10-layered
GeoTIFF) and can only be
returned by the function when it is already present as a file.
See R-code in the \href{https://github.com/inbo/n2khab-inputs}{
n2khab-inputs} repository for its creation from
the \code{GRTSmaster_habitats} data source.

Both GeoTIFFs (\code{GRTSmaster_habitats}, \code{GRTSmh_brick}) use the
\code{INT4S} datatype.

The higher-level ranking numbers of the RasterBrick allow spatially balanced
samples at lower spatial resolution than that of 32 m, and can also be
used for aggregation purposes.
The provided hierarchical levels correspond to the resolutions vector
\code{32 * 2^(0:9)} (minimum: 32 meters, maximum: 16384 meters), with
the corresponding RasterBrick layers named as \code{level0} to \code{level9}.
}
\examples{
\dontrun{
# This example supposes that your working directory is a folder next to
# the 'n2khab-inputs' repository AND that the
# 'GRTSmaster_habitats' data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can consider
# what to do.
r <- read_GRTSmh("../n2khab-inputs/data")
r
}

}
\references{
Stevens D.L. & Olsen A.R. (2003). Variance estimation for spatially balanced
samples of environmental resources. Environmetrics 14 (6): 593–610.
DOI: 10.1002/env.606.

Stevens D.L. & Olsen A.R. (2004). Spatially Balanced Sampling of Natural
Resources. Journal of the American Statistical Association 99 (465):
262–278. DOI: 10.1198/016214504000000250.
}
\seealso{
Other functions involved in processing the \code{GRTSmaster_habitats} data source: \code{\link{base4frac_to_dec}},
  \code{\link{dec_to_base4frac}},
  \code{\link{read_GRTSmh_base4frac}},
  \code{\link{read_GRTSmh_diffres}}
}
\concept{functions involved in processing the \code{GRTSmaster_habitats} data source}
