% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatmap}
\alias{read_habitatmap}
\title{Return the data source \code{habitatmap}}
\usage{
read_habitatmap(path = fileman_up("n2khab_data"),
  file = "10_raw/habitatmap")
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, use this
argument in scripts to set the location of the folder '\strong{\code{n2khab_data}}'.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing} repository.}
}
\value{
An sf object
}
\description{
\code{read_habitatmap} returns all polygons of the \code{habitatmap} (De Saeger et al., 2018) that (partially)
contain habitat or a Regional Important Biotope (RIB).
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatmapd'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

r <- read_habitatmap()
}

}
\references{
De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P.,
Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische
Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het
Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO).
DOI: https://doi.org/10.21436/inbor.15138099.
}
\seealso{
Other functions involved in processing the \code{habitatmap} data source: \code{\link{read_habitatmap_stdized}},
  \code{\link{read_watersurfaces_hab}}
}
\concept{functions involved in processing the \code{habitatmap} data source}
