% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleframe_habitat.R
\name{read_habitatmap}
\alias{read_habitatmap}
\title{Return the data source \code{habitatmap_processed}}
\usage{
read_habitatmap(path,
  file = "20_processed/habitatmap/habitatmap_processed.gpkg")
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, use this argument
in scripts to set the location of the folder '\strong{\code{data}}'.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{https://github.com/inbo/n2khab-inputs}{n2khab-inputs} repository.}
}
\value{
A list with following items:
  \itemize{
  \item \code{polygons}: an sf object with two variables
  \itemize{
    \item \code{polygon_id}
    \item \code{polygon_description}}
  }
  \itemize{
  \item \code{patches}: a dataframe with following variables
  \itemize{
    \item \code{polygon_id}
    \item \code{patch_id}
    \item \code{code_orig}: original vegetation code in \code{habitat_map}.
    \item \code{phab}: proportion of polygon covered by patch.
    \item \code{uncertain}: \code{TRUE} when vegetation type is not certain.
    \item \code{type}: habitat or RIB type listed in \code{\link{types}}.
    \item \code{typelevel}: see \code{\link{types}}.
    \item \code{main_type}: see \code{\link{types}}.
    \item \code{typeclass}: see \code{\link{types}}.}
    }
}
\description{
The data source \code{habitatmap_processed} is a GeoPackage that contains two layers:
\code{habitatmap_polygons} and \code{habitatmap_patches}.
}
\details{
The \code{habitatmap_polygons} layer consists of all polygons of the raw
\code{habitatmap} data source (De Saeger et al., 2018) that contain habitat or RIB
which are listed in \code{\link{types}}. Every polygon in the \code{habitatmap}
can consist of maximum 5 different vegetation types. This information is stored
in the columns 'HAB1', 'HAB2',..., 'HAB5' of the attribute table. The fraction
of each vegetation type within the polygons is stored in the columns 'PHAB1',
'PHAB2', ..., 'PHAB5'. These polygon fractions are called 'patches'.

In the \code{habitatmap_patches} layer, every row corresponds with one patch. A new variable
is created to provide the patch number (1 to 5): 'patch_id'.

For some patches the vegetation type is uncertain, and the vegetation code consists
of 2 or 3 possible types, separated with a ','. In the \code{habitatmap_patches} layer,
the different possible vegetation types are split up and one row is created for each of
them. The variable \code{uncertain} will be \code{TRUE} if a patch consists of 2 or 3
possible vegetation types, and \code{FALSE} if only one vegetation type is provided.

For some patches the original vegetation code in the \code{habitatmap} was not consistent
with the general coding rules. In that case the code was corrected. The file
\code{habitatmap_correction} gives an overview of the adjusted codes.
}
\examples{
\dontrun{
# This example supposes that your working directory is a folder next to
# the 'n2khab-inputs' repository AND that the
# 'habitatmap_processed' data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can consider
# what to do.
r <- read_habitatmap("../n2khab-inputs/data")
r_polygons <- r$polygons
r_patches <- r$patches
}

}
\references{
De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P.,
Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische
Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het
Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO).
DOI: 10.21436/inbor.15138099.
}
\concept{functions involved in processing the \code{habitat_map} data source}
