% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatmap}
\alias{read_habitatmap}
\title{Return the data source \code{habitatmap} as an \code{sf} multipolygon layer}
\usage{
read_habitatmap(
  file = file.path(fileman_up("n2khab_data"), "10_raw/habitatmap"),
  filter_hab = FALSE
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{filter_hab}{If \code{TRUE} only polygons that (partially) contain habitat or a regionally
important biotope (RIB) are returned. The default value is \code{FALSE}.}
}
\value{
A Simple feature collection of
type \code{MULTIPOLYGON}.
}
\description{
Returns the raw data source \code{habitatmap} (De Saeger et al., 2018)
as a standardized \code{sf} multipolygon layer
(tidyverse-styled, internationalized) in the Belgian Lambert 72 CRS
(EPSG-code \href{https://epsg.io/31370}{31370}).
Given the size of the data source, this function
takes a bit longer than usual to run.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'habitatmap'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hm <- read_habitatmap()
hm
}

}
\references{
De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P.,
Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische
Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het
Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO).
\doi{10.21436/inbor.15138099}.
}
\seealso{
Other functions involved in processing the 'habitatmap' data source: 
\code{\link{read_habitatmap_stdized}()},
\code{\link{read_habitatmap_terr}()},
\code{\link{read_watersurfaces_hab}()}
}
\concept{functions involved in processing the 'habitatmap' data source}
