% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatmap_stdized}
\alias{read_habitatmap_stdized}
\title{Return the data source \code{habitatmap_stdized} as a list of two
objects}
\usage{
read_habitatmap_stdized(path,
  file = "20_processed/habitatmap_stdized/habitatmap_stdized.gpkg")
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, use this
argument in scripts to set the location of the folder '\strong{\code{data}}'.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing} repository.}
}
\value{
A list of two objects:
  \itemize{
  \item \code{habitatmap_polygons}: an sf object of \code{habitatmap} polygons with two attribute variables
  \itemize{
    \item \code{polygon_id}
    \item \code{description_orig}: polygon description based on the
    orginal vegetation codes in the raw \code{habitatmap}}
  }
  \itemize{
  \item \code{habitatmap_patches}: a tibble with following variables
  \itemize{
    \item \code{polygon_id}
    \item \code{patch_id}
    \item \code{code_orig}: original vegetation code in raw \code{habitatmap}.
    \item \code{phab}: proportion of polygon covered by patch, as a percentage.
    \item \code{certain}: \code{TRUE} when vegetation type is certain and
     \code{FALSE} when vegetation type is uncertain.
    \item \code{type}: habitat or RIB type listed in \code{\link{types}}.
    }
    }
}
\description{
\code{read_habitatmap_stdized} returns the data source \code{habitatmap_stdized} as a list of two objects:
\itemize{
  \item \code{habitatmap_polygons}: an sf object with all polygons
  of the \code{habitatmap} that contain habitat or a Regionally
  Important Biotope (RIB).
  \item \code{habitatmap_patches}: a tibble with information on the
  habitat and RIB patches (HAB1, HAB2,..., HAB5) that occur within
  the \code{habitatmap_polygons}, each row corresponding with one
  patch.
  }
}
\details{
The data source \code{habitatmap_stdized} is the processed version
of the raw data source \code{habitatmap} (De Saeger et al., 2018).
Every polygon in the \code{habitatmap} can consist of maximum 5
different vegetation types. This information is stored in the
columns 'HAB1', HAB2',..., 'HAB5' of the attribute table. The
fraction of each vegetation type within the polygons is stored in
the columns 'PHAB1', 'PHAB2', ..., 'PHAB5'. The different parts of
the polygons are called 'patches'.

The data source \code{habitatmap_stdized} is a GeoPackage that
contains:
\itemize{
  \item \code{habitatmap_polygons}: a spatial layer with every
  \code{habitatmap} polygon that contains a habitat or RIB type
  listed in \code{\link{types}}.
  \item \code{habitatmap_patches}: a table in which every row
  corresponds with one patch.
  }

The processing of the \code{habitatmap_patches} tibble included
following steps:
\itemize{
  \item A new variable was created to provide the
  patch number (1 to 5): 'patch_id'.
  \item For some patches the vegetation type is uncertain, and the
  vegetation code in the raw \code{habitatmap} data source consists
  of 2 or 3 possible types, separated with a ','. The different
  possible vegetation types are split up and one row is created for
  each of them. The variable \code{certain} will be \code{FALSE} if
  a patch consists of 2 or 3 possible vegetation types, and \code{TRUE}
  if only one vegetation type is provided.
  \item For some patches the original vegetation code in the
  \code{habitatmap} was not consistent with general coding syntax or
  with the type codes from the \code{\link{types}}. In that case the
  code was adjusted.

}

The R-code for creating the \code{habitatmap_stdized} data source
  can be found in the \href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing} repository.
}
\examples{
\dontrun{
# This example supposes that your working directory is a folder next
# to the 'n2khab-preprocessing' repository AND that the 'habitatmap_stdized
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

r <- read_habitatmap("../n2khab-preprocessing/data")
r_polygons <- r$habitatmap_polygons
r_patches <- r$habitatmap_patches
}

}
\references{
De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P.,
Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische
Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het
Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO).
DOI: 10.21436/inbor.15138099.
}
\concept{functions involved in processing the \code{habitatmap} data source}
