% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatmap_terr}
\alias{read_habitatmap_terr}
\title{Return the data source \code{habitatmap_terr} as a list of two
objects}
\usage{
read_habitatmap_terr(path = fileman_up("n2khab_data"),
  file = "20_processed/habitatmap_terr/habitatmap_terr.gpkg",
  keep_aq_patches = TRUE)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{../doc/v020_datastorage.html}{vignette} on data storage.}

\item{keep_aq_patches}{Logical; \code{TRUE} by default.
The data source \code{habitatmap_terr} aims at delineating all
polygons with at least one (semi-)terrestrial type (i.e. patch).
For those polygons, it returns all known habitat types and RIBs as patches.
Hence, in several cases polygons do occur with a combination of terrestrial
and aquatic patches.
Setting \code{keep_aq_patches = FALSE} is convenient for use cases where one
only wants to look at the (semi-)terrestrial patches: this setting will
discard all aquatic patches in 'mixed' aquatic/terrestrial polygons.
The aquatic types are defined as those with hydrological class \code{"HC3"}
(see \code{tag_2}) in the \code{\link{types}} data source.}
}
\value{
A list of two objects:
  \itemize{
  \item \code{habitatmap_terr_polygons}: a Simple feature collection of
  geometry type \code{MULTIPOLYGON} with four attribute variables:
  \itemize{
    \item \code{polygon_id}
    \item \code{description_orig}: polygon description based on the
    original vegetation codes in the \code{habitatmap} data source
    \item \code{description}: based on \code{description_orig} but with the
    interpreted type codes
    \item \code{source}: states where \code{description} comes from: either
    \code{habitatmap_stdized} or \code{habitatmap_stdized + interpretation}
  }
  \item \code{habitatmap_terr_patches}: a tibble with the following
  variables (the first 5 being identical to those in
  \code{habitatmap_stdized}):
  \itemize{
    \item \code{polygon_id}
    \item \code{patch_id}
    \item \code{code_orig}
    \item \code{phab}
    \item \code{certain}
    \item \code{type}: the interpreted habitat or RIB type
    \item \code{source}: states where \code{type} comes from: either
    \code{habitatmap_stdized} or \code{habitatmap_stdized + interpretation}
    }
    }
}
\description{
\code{read_habitatmap_terr()} returns the data source \code{habitatmap_terr}
as a list of two objects: \code{habitatmap_terr_polygons} and
\code{habitatmap_terr_patches}.
\code{habitatmap_terr} is the further interpreted, terrestrial part of
\code{habitatmap_stdized} (see \code{\link{read_habitatmap_stdized}}).
}
\details{
\code{habitatmap_terr} was derived from \code{habitatmap_stdized} as
follows:
\itemize{
\item{it excludes all polygons
that are most probably aquatic habitat or RIB.
Also \code{1130} is excluded.
In the process, a distinction is also made between \code{2190_a} and
\code{2190_overig};}
\item{it excludes patches which most probably are \emph{no}
habitat or RIB at all.
Those are the patches where \code{code_orig} contains \code{"bos"} or is
equal to \code{“6510,gh”} or \code{“9120,gh”};}
\item{it translates several main type codes into a corresponding
subtype which they almost always represent:
\code{6410} -> \code{6410_mo},
\code{6430} -> \code{6430_hf},
\code{6510} -> \code{6510_hu},
\code{7140} -> \code{7140_meso},
\code{9130} -> \code{9130_end};}
\item{it distinguishes types \code{rbbhfl} and \code{rbbhf}.}
}

The data source \code{habitatmap_terr} is a GeoPackage, available at
\href{https://doi.org/10.5281/zenodo.3468948}{Zenodo}, that contains:
\itemize{
  \item{\code{habitatmap_terr_polygons}: a spatial polygon layer}
  \item{\code{habitatmap_terr_patches}: a table in which every row
  corresponds with one patch.}
  }

The R-code for creating the \code{habitatmap_terr} data source
can be found in the
\href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing}
repository.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatmap_terr'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

habmap_terr <- read_habitatmap_terr()
habmap_terr$habitatmap_terr_polygons
habmap_terr$habitatmap_terr_patches

habmap_terr_noaq <- read_habitatmap_terr(keep_aq_patches = FALSE)
habmap_terr_noaq$habitatmap_terr_polygons
habmap_terr_noaq$habitatmap_terr_patches
}

}
\seealso{
Other functions involved in processing the \code{habitatmap} data source: \code{\link{read_habitatmap_stdized}},
  \code{\link{read_habitatmap}},
  \code{\link{read_watersurfaces_hab}}
}
\concept{functions involved in processing the \code{habitatmap} data source}
