% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatsprings}
\alias{read_habitatsprings}
\title{Return the data source \code{habitatsprings} as an \code{sf} point
layer}
\usage{
read_habitatsprings(
  path = fileman_up("n2khab_data"),
  file = "10_raw/habitatsprings/habitatsprings.geojson"
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}
}
\value{
A Simple feature collection of
type \code{POINT}, with attribute variables:
\itemize{
\item \code{point_id}
\item \code{name}: site name.
\item \code{code_orig}: original vegetation code in raw
\code{habitatsprings}.
\item \code{type}: habitat type listed in \code{\link{types}}.
\item \code{certain}: \code{TRUE} when vegetation type is certain and
\code{FALSE} when vegetation type is uncertain.
\item \code{area_m2}: area as square meters.
\item \code{year}: year of field inventory.
\item \code{in_sac}: logical.
Is the site situated within a Special Area of Conservation?
\item \code{source}: original data source of the record.
}

Note that the \code{type} variable has implicit \code{NA} values in this
case
(i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can add
it as a level with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.
}
\description{
Returns the raw data source \code{habitatsprings} as an \code{sf} point
layer in the Belgian Lambert 72 CRS (EPSG-code
\href{https://epsg.io/31370}{31370}).
}
\details{
The data source \code{habitatsprings} is a GeoJSON file (conforming to
the RFC7946 specifications), available at
\href{https://doi.org/10.5281/zenodo.3550994}{Zenodo}.
It represents sites that correspond with presence or absence
of the Natura 2000 habitat type \code{7220} (Petrifying springs with tufa
formation (\emph{Cratoneurion})) in springs and streaming water segments in
the Flemish Region, Belgium.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatsprings'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hs <- read_habitatsprings()
hs
}

}
