% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_habitatstreams}
\alias{read_habitatstreams}
\title{Return the data source \code{habitatstreams} as an \code{sf} linestring
layer or as a list}
\usage{
read_habitatstreams(
  path = fileman_up("n2khab_data"),
  file = "10_raw/habitatstreams",
  source_text = FALSE
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}

\item{source_text}{Logical, defaults to \code{FALSE}.
If \code{TRUE}, the list version is returned (see \emph{Value}).}
}
\value{
With \code{source_text = FALSE} (default): a Simple feature collection of
type \code{LINESTRING}.

With \code{source_text = TRUE}: a list of two objects:
\itemize{
\item \code{lines}: the same \code{sf} object as with \code{source_text = FALSE}.
\item \code{sources}: textual explanation on the values of the \code{source_id}
variable in the \code{sf} object.
}
}
\description{
Returns the raw data source \code{habitatstreams} (Leyssen et al., 2018)
as an \code{sf} linestring
layer or as a list of two objects: the \code{sf} object plus a data frame
with textual explanation of the values of the \code{source_id} variable.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatstreams'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

library(magrittr)
library(sf)
hs <- read_habitatstreams()
hs
hs2 <- read_habitatstreams(source_text = TRUE)
hs2
all.equal(hs \%>\% st_drop_geometry,
          hs2$lines \%>\% st_drop_geometry)
}

}
\references{
Leyssen A., Denys L. De Saeger S. (2018). Indicatieve situering van het
Natura 2000 habitattype 3260. Submontane en laaglandrivieren met vegetaties
behorend tot het \emph{Ranunculion fluitantis} en het
\emph{Callitricho-Batrachion}.
Uitgave 2018 (versie 1.6). Rapporten van het Instituut voor Natuur- en
Bosonderzoek 2018 (72). Research Institute for Nature and Forest, Brussels.
DOI: https://doi.org/10.21436/inbor.15138370
}
