% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shallowgroundwater.R
\name{read_shallowgroundwater}
\alias{read_shallowgroundwater}
\title{Return the data source \code{shallowgroundwater} as an \code{sf}
multipolygon layer}
\usage{
read_shallowgroundwater(
  file = file.path(fileman_up("n2khab_data"),
    "10_raw/shallowgroundwater/shallowgroundwater.gpkg")
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns the raw data source \code{shallowgroundwater}
as an \code{sf} multipolygon
layer.
The coordinate reference system is 'Belge 72 / Belgian Lambert 72'
(EPSG-code \href{https://epsg.io/31370}{31370}).
}
\details{
The R code for creating the \code{shallowgroundwater} data source,
starting from a manually composed layer, can be found in the
\href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing}
repository.
}
\examples{
\dontrun{
shallowgroundwater <- read_shallowgroundwater()
shallowgroundwater
}

}
