% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilmap.R
\name{read_soilmap}
\alias{read_soilmap}
\title{Return the \code{soilmap} or \code{soilmap_simple} data source as an
\code{sf} object}
\usage{
read_soilmap(path = fileman_up("n2khab_data"), file = "10_raw/soilmap",
  use_processed = TRUE, standardize_polders = FALSE,
  simplify = FALSE)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{../doc/v020_datastorage.html}{vignette} on data storage.}

\item{use_processed}{Load and return the processed data source
\code{soilmap_simple}, instead of the raw data source \code{soilmap}.}

\item{standardize_polders}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, fill empty values of substrate,
texture and moisture in the 'Polders' area with translated values from
De Vos et al XXXX.
Also, use the value of \code{soiltype_unified} provided by this source.}

\item{simplify}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, only return a limited number of variables.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns either the raw data source \code{soilmap} or (by default) the
processed data source \code{soilmap_simple}.
}
\details{
In case the raw data source \code{soilmap} is used, it is possible to
manually perform the standardization (for polders) and/or the simplification,
which were both applied in the \code{soilmap_simple} data source.

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation of the
\code{soilmap_simple} data source from
the \code{soilmap} data source.

TO BE ADDED: explanation on the variables in the result.
}
