% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilmap.R
\name{read_soilmap}
\alias{read_soilmap}
\title{Return the \code{soilmap} or \code{soilmap_simple} data source as an
\code{sf} object}
\usage{
read_soilmap(
  path = fileman_up("n2khab_data"),
  file = "20_processed/soilmap_simple/soilmap_simple.gpkg",
  file_raw = "10_raw/soilmap",
  use_processed = TRUE,
  standardize_coastalplain = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the \emph{processed} data source \code{soilmap_simple}.
May include a path prefix.
Used only if \code{use_processed = TRUE} (= default).
The default value follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}

\item{file_raw}{Same as \code{file}, to define the name (subpath) of the \emph{raw}
datasource \code{soilmap}.
Used only if \code{use_processed = FALSE}.}

\item{use_processed}{Logical.
If \code{TRUE} (the default), load and return the processed data source
\code{soilmap_simple}, instead of the raw data source \code{soilmap}.
\code{soilmap_simple} can be manually generated by reading the raw \code{soilmap} data
source with \code{standardize_coastalplain=TRUE} and \code{simplify=TRUE}, but this
takes some time.}

\item{standardize_coastalplain}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, fill the values of the morphogenetic substrate,
texture and drainage variables (\code{bsm_mo_substr}, \code{bsm_mo_tex} and
\code{bsm_mo_drain} + their \verb{_explan} counterparts) \emph{for features in the
'coastalplain' area}.
\itemize{
\item To derive morphogenetic texture and drainage levels from the geomorphological
ones, a translation table by Bruno De Vos &
Carole Ampe is applied (for earlier work on this, see Ampe 2013).
\item Substrate classes are copied over from \code{bsm_ge_substr} into \code{bsm_mo_substr}
(\code{bsm_ge_substr} already followed the categories of \code{bsm_mo_substr}).
}

These steps coincide with the approach that was taken to construct
\code{bsm_mo_soilunitype} in the raw data source.}

\item{simplify}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, only return a limited number of variables.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns either the raw data source \code{soilmap}
or (by default) the
processed data source \code{soilmap_simple}.
}
\details{
The raw data source is published
\href{https://www.dov.vlaanderen.be/geonetwork/srv/dut/catalog.search#/metadata/5c129f2d-4498-4bc3-8860-01cb2d513f8f}{at DOV}
(Databank Ondergrond Vlaanderen) and
is discussed by Van Ranst & Sys (2000) and Dudal et al. (2005).
A 'pure' (single) dataformat of the raw data source (no metadatafiles etc.)
has also been stored (with versioning) at
Zenodo (\url{https://doi.org/10.5281/zenodo.3387007}) - which we refer to as the
\code{soilmap} data source - in order
to support the \code{read_soilmap()} function and to sustain long-term workflow
reproducibility.

Note that factors are generated with implicit \code{NA} values (i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can convert
such variables with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.

In case the raw data source \code{soilmap} is used
(\code{use_processed = FALSE}), it is possible to
manually perform the standardization for coastal plain features and/or the
simplification,
both of which were applied in the \code{soilmap_simple} data source.
See \emph{Arguments} for more information.

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation of the
\code{soilmap_simple} data source from
the \code{soilmap} data source.
}
\references{
\itemize{
\item Ampe C. (2013). Databank aardewerk Vlaanderen 2010.
Omzetten (zeer) oude legende bodemkartering naar legende bodemkaart
Kuststreek. Vlaamse Landmaatschappij Regio West, Bruges, 45 p.
\item Dudal R., Deckers J., Van Orshoven J. & Van Ranst E. (2005). Soil survey
in Belgium and its applications. In: Bullock P., Jones R.J.A., Montanarella
L. (editors). Soil Resources of Europe. Office for Official Publications of
the European Communities, Luxembourg, p. 63–71.
URL: http://hdl.handle.net/1854/LU-368514.
\item Van Ranst E. & Sys C. (2000). Eenduidige legende van de digitale
bodemkaart van Vlaanderen (schaal 1: 20000). Universiteit Gent, Laboratorium
voor Bodemkunde, Ghent, 361 p. URL: http://hdl.handle.net/1854/LU-125899.
}
}
