% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilmap.R
\name{read_soilmap}
\alias{read_soilmap}
\title{Return the \code{soilmap} or \code{soilmap_simple} data source as an
\code{sf} object}
\usage{
read_soilmap(
  path = fileman_up("n2khab_data"),
  file = "10_raw/soilmap",
  use_processed = TRUE,
  standardize_coastalplain = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}

\item{use_processed}{Logical.
If \code{TRUE} (the default), load and return the processed data source
\code{soilmap_simple}, instead of the raw data source \code{soilmap}.
\code{soilmap_simple} can be manually generated by reading the raw \code{soilmap} data
source with \code{standardize_coastalplain=TRUE} and \code{simplify=TRUE}, but this
takes some time.}

\item{standardize_coastalplain}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, fill the values of the morphogenetic substrate,
texture and drainage variables (\code{bsm_mo_substr}, \code{bsm_mo_tex} and
\code{bsm_mo_drain} + their \verb{_explan} counterparts) \emph{for features in the
'coastalplain' area} (where these are missing).
To derive morphogenetic texture and drainage levels from the geomorphological
ones, a translation table by Bruno De Vos &
Carole Ampe is applied (for earlier work on this, see Ampe 2013).
Substrate classes are copied over from \code{bsm_ge_substr} into \code{bsm_mo_substr}
(hence, the resulting \code{bsm_mo_substr} is actually hybrid: 'mo' + 'ge').
These steps coincide with the approach that was taken to construct
\code{bsm_mo_soilunitype} in the raw data source.}

\item{simplify}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, only return a limited number of variables.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns either the raw data source \code{soilmap} or (by default) the
processed data source \code{soilmap_simple}.
}
\details{
Note that factors are generated with implicit \code{NA} values (i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can convert with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.

In case the raw data source \code{soilmap} is used, it is possible to
manually perform the standardization for coastal plain features and/or the
simplification,
both of which were applied in the \code{soilmap_simple} data source
See \emph{Arguments} for more information.

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation of the
\code{soilmap_simple} data source from
the \code{soilmap} data source.
}
\references{
Ampe C. (2013). Databank aardewerk Vlaanderen 2010.
Omzetten (zeer) oude legende bodemkartering naar legende bodemkaart
Kuststreek. Vlaamse Landmaatschappij Regio West, Bruges, 45 p.
}
