% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soilmap.R
\name{read_soilmap}
\alias{read_soilmap}
\title{Return the \code{soilmap} or \code{soilmap_simple} data source as an
\code{sf} object}
\usage{
read_soilmap(
  path = fileman_up("n2khab_data"),
  file = "10_raw/soilmap",
  use_processed = TRUE,
  standardize_coastalplain = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}

\item{use_processed}{Logical.
If \code{TRUE}, load and return the processed data source
\code{soilmap_simple}, instead of the raw data source \code{soilmap}.}

\item{standardize_coastalplain}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, fill empty values of substrate,
texture and drainage classes in the 'coastalplain' area.
For texture and drainage classes, a translation table by Bruno De Vos &
Carole Ampe is applied (for earlier work on this, see Ampe 2013).}

\item{simplify}{Logical.
Only applied with \code{use_processed = FALSE}.
If \code{TRUE}, only return a limited number of variables.}
}
\value{
A Simple feature collection of geometry type \code{MULTIPOLYGON}.
}
\description{
Returns either the raw data source \code{soilmap} or (by default) the
processed data source \code{soilmap_simple}.
}
\details{
Note that factors are generated with implicit \code{NA} values (i.e. there is
no factor level to represent the missing values).
If you want this category to appear in certain results, you can convert with
\code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}.

In case the raw data source \code{soilmap} is used, it is possible to
manually perform the standardization (for coastalplain) and/or the simplification,
which were both applied in the \code{soilmap_simple} data source.

See R-code in the \href{https://github.com/inbo/n2khab-preprocessing}{
n2khab-preprocessing} repository for the creation of the
\code{soilmap_simple} data source from
the \code{soilmap} data source.

TO BE ADDED: explanation on the variables in the result.
}
\references{
Ampe C. (2013). Databank aardewerk Vlaanderen 2010.
Omzetten (zeer) oude legende bodemkartering naar legende bodemkaart
Kuststreek. Vlaamse Landmaatschappij Regio West, Bruges, 45 p.
}
