% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_textdata.R
\name{read_types}
\alias{read_types}
\title{Return the 'types' data source as a tibble with names & shortnames}
\usage{
read_types(path = pkgdatasource_path("textdata/types", ".tsv"),
  file = "types", file_namelist = "namelist", lang = "en")
}
\arguments{
\item{path}{Location of the data sources \code{types} and \code{namelist}.
The default is to use the location of the data sources as delivered by
the installed package.}

\item{file}{The filename of the \code{types} data source, without extension.
The default is to use the file delivered by the installed package.}

\item{file_namelist}{The filename of the \code{namelist} data source,
without extension.
The default is to use the file delivered by the installed package.}

\item{lang}{An
\href{https://www.w3.org/International/articles/language-tags/index.en}{
IETF BCP 47 language tag}, such as \code{"en"} or \code{"nl"}, to specify
the language of names & shortnames to be returned in the tibble.}
}
\value{
The \code{types} dataframe as a \code{\link[tibble:tbl_df-class]{tibble}},
with names & shortnames added for types, typeclasses and tags
according to the \code{lang} argument.
The tibble has 108 rows and 16 variables.
See \code{\link{types}} for documentation of the data-source's contents.
See \code{\link{namelist}} for the link between codes or other identifiers
and the corresponding names (and shortnames).

The added names and shortnames are represented by the following variables:
\itemize{
  \item \code{type_name}
  \item \code{type_shortname}
  \item \code{typeclass_name}: a factor with the level order
  according to that of \code{typeclass}.
  \item \code{tag_1_name}
  \item \code{tag_1_shortname}
  \item \code{tag_2_name}
  \item \code{tag_2_shortname}
  \item \code{tag_3_name}
  \item \code{tag_3_shortname}
}
}
\description{
Returns the included data source \code{\link{types}} as a
\code{\link[tibble:tbl_df-class]{tibble}}.
Names and shortnames from \code{\link{namelist}} are added,
in English by default.
}
\details{
\code{\link{types}} is a data source in the
\href{https://inbo.github.io/git2rdata/index.html}{vc-format} which provides
a checklist of types, represented by their \strong{current} codes, together
with several attributes.
A 'type' refers to either a (main) habitat type, a
habitat subtype or a regionally important biotope (RIB).

\code{read_types()} reads the \code{\link{types}} data source, adds
names + shortnames
and returns it as a
\code{\link[tibble:tbl_df-class]{tibble}}.
A tibble is a dataframe that makes working in the tidyverse a little
\href{https://r4ds.had.co.nz/tibbles.html}{easier}.
By default, the data version delivered with the package is used and English
names (\code{lang = "en"}) are returned for types, typeclasses and tags.
}
\section{Recommended usage}{


  \code{read_types()}

  \code{read_types(lang = "nl")}
}

\examples{
\dontrun{
read_types()
read_types(lang = "nl")
}

}
\seealso{
\code{\link{types}}

Other reading functions for n2khab-referencelists: \code{\link{read_env_pressures}},
  \code{\link{read_namelist}}, \code{\link{read_schemes}}
}
\concept{reading functions for n2khab-referencelists}
