% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_watersurfaces}
\alias{read_watersurfaces}
\title{Return the data source \code{watersurfaces} as an \code{sf} polygon layer}
\usage{
read_watersurfaces(
  path = fileman_up("n2khab_data"),
  file = "10_raw/watersurfaces",
  extended = FALSE
)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).}

\item{extended}{Logical.
Should names or explanations of codes be added as extra
variables in the result?
Currently only applies to \code{wfd_type}; if \code{TRUE}, a variable
\code{wfd_type_name} is added.
Defaults to \code{FALSE}.}
}
\value{
A Simple feature collection of
type \code{POLYGON}, sorted by \code{polygon_id}, with the following
variables (not mentioning extra 'name' variables for
\code{extended = TRUE}):
\itemize{
  \item \code{polygon_id}: code of the polygon;
  \item \code{wfd_code}: optional; Flemish code of the water body with
  regard to the Water Framework Directive (WFD);
  \item \code{hyla_code}: optional; code of the watersurface according to the
  Flemish working group 'Hyla', a working group on amphibians & reptiles;
  \item \code{name}: name of the watersurface;
  \item \code{area_name}: name of the area;
  \item \code{wfd_type}: type code according to the Flemish WFD typology
  (Denys, 2009);
  \item \code{wfd_type_certain}: Logical.
  Is there high confidence about the \code{wfd_type} determination?
  \item \code{depth_class}: class of water depth;
  \item \code{connectivity}: connectivity class;
  \item \code{usage}: usage class.
}
}
\description{
Returns the raw data source \code{watersurfaces} (Packet et al., 2018)
as a standardized \code{sf} polygon layer
(tidyverse-styled, internationalized).
}
\details{
See Packet et al. (2018) for an elaborate explanation of the data source
and its attributes.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'watersurfaces'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

ws <- read_watersurfaces()
ws
summary(ws)

ws2 <- read_watersurfaces(extended = TRUE)
ws2
}

}
\references{
\itemize{
\item Denys L. (2009). Een a posteriori typologie van stilstaande, zoete
wateren in Vlaanderen.
Rapporten van het Instituut voor Natuur- en Bosonderzoek INBO.R.2009.34.
Instituut voor Natuur- en Bosonderzoek, Brussel.
\item Packet J., Scheers K., Smeekens V., Leyssen A., Wils C. & Denys L.
(2018).
Watervlakken versie 1.0: polygonenkaart van stilstaand water in Vlaanderen.
Een nieuw instrument voor onderzoek, water-, milieu- en natuurbeleid.
Rapporten van het Instituut voor Natuur- en Bosonderzoek 2018 (14).
Instituut voor Natuur- en Bosonderzoek, Brussel.
DOI: \url{https://doi.org/10.21436/inbor.14178464}.
}
}
\seealso{
Other functions involved in processing the \code{watersurfaces} data source: 
\code{\link{read_watersurfaces_hab}()}
}
\concept{functions involved in processing the \code{watersurfaces} data source}
\concept{functions returning important environmental data sets}
