% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_watersurfaces_hab}
\alias{read_watersurfaces_hab}
\title{Return the data source \code{watersurfaces_hab} as a list of two
objects}
\usage{
read_watersurfaces_hab(
  file = file.path(locate_n2khab_data(),
    "20_processed/watersurfaces_hab/watersurfaces_hab.gpkg"),
  interpreted = FALSE,
  collapse = TRUE,
  version = c("watersurfaces_hab_v6", "watersurfaces_hab_v5", "watersurfaces_hab_v4",
    "watersurfaces_hab_v3", "watersurfaces_hab_v2", "watersurfaces_hab_v1",
    "watersurfaces_hab_v6.1_interim")
)
}
\arguments{
\item{file}{The absolute or relative file path of the data source.
The default follows the data management advice in the
vignette on data storage (run \code{vignette("v020_datastorage")}).
It uses the first \code{n2khab_data} folder that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{interpreted}{If \code{TRUE}, the interpreted subtype is provided when the subtype is missing. This only
applies to type 3130. When the subtype is missing for 3130, we interpret it as 3130_aom.}

\item{collapse}{Logical.
Should the resulting \code{watersurfaces_types} list element have a single
row for each combination of \code{polygon_id} and \code{type}?
This causes collapsing:
\itemize{
  \item as a single string of different values of \code{code_orig} that led
  to the same \code{type};
  \item of values of \code{certain}, using the \code{\link{any}} summary
  function.
}}

\item{version}{Version ID of the data source.
Defaults to the latest available version defined by the package.
Versions with the 'interim' suffix are designed to be used within the Research
Institute for Nature and Forest (INBO) only.}
}
\value{
A list of two objects:
  \itemize{
  \item \code{watersurfaces_polygons}: an sf object of standing water polygons with four attribute variables:
  \itemize{
    \item \code{polygon_id}
    \item \code{polygon_id_ws}: id for the polygon in the \code{watersurface map}
    \item \code{polygon_id_habitatmap}: id's of all overlapping polygons of \code{habitatmap_stdized} that
    contain standing water habitat. The different id's are separated by '+'.
    \item \code{description_orig}: descriptions of all overlapping polygons of \code{habitatmap_stdized} that
    contain standing water habitat. The different descriptions are separated by '+'.}
  }
  \itemize{
  \item \code{watersurfaces_types}: a tibble with following variables:
  \itemize{
    \item \code{polygon_id}
    \item \code{type}: habitat or RIB type listed in \code{\link{types}}.
    \item \code{certain}: \code{TRUE} when the type is certain and
     \code{FALSE} when the type is uncertain.
    \item \code{code_orig}: original type code in raw \code{habitatmap}.
    }
    }
}
\description{
\code{read_watersurfaces_hab} returns the data source \code{watersurfaces_hab} as a list of two objects:
\itemize{
  \item \code{watersurfaces_polygons}: an sf object in the Belgian Lambert 72
  CRS (EPSG-code \href{https://epsg.io/31370}{31370}) with all polygons
  that contain standing water types (habitat or RIB).
  \item \code{watersurfaces_types}: a tibble with information on the
  standing water \link{types} (HAB1, HAB2,..., HAB5) that occur within
  each polygon of \code{watersurfaces_polygons}.
  }
}
\details{
The data source \code{watersurfaces_hab} is a combination of \code{habitatmap_stdized} (see
\code{\link{read_habitatmap_stdized}}) and the
\href{https://doi.org/10.5281/zenodo.3386857}{watersurface map of Flanders}.
It contains all standing water types in Flanders.


The data source \code{watersurfaces_hab} is a GeoPackage, available at
\href{https://doi.org/10.5281/zenodo.3374645}{Zenodo}, that contains:
\itemize{
  \item \code{watersurfaces_hab_polygons}: a spatial layer with all polygons that contain standing water types listed in \code{\link{types}}.
  \item \code{watersurfaces_hab_types}: a table in which every row corresponds with a combination of polygon and type.
  }

The polygons with 2190_a habitat (dune slack ponds) are generated by selecting all watersurface polygons that
overlap with dune habitat polygons (21xx) of the standardized habitat map.

For each of the other considered habitat types (31xx and rbbah) we select the watersurface polygons that
overlap with the selected habitat type polygons of the standardized habitat map. We also select polygons of the
standardized habitat map that contain standing water types but do not overlap with any watersurface polygon of the
watersurface map.

The R-code for creating the \code{watersurfaces_hab} data source can be found in the \href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing}
repository.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'watersurfaces_hab'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

wsh <- read_watersurfaces_hab()
wsh_polygons <- wsh$watersurfaces_polygons
wsh_types <- wsh$watersurfaces_types
wsh_polygons
wsh_types
}

}
\references{
\itemize{
\item Leyssen A., Scheers K., Packet J., Van Hecke F., Wils C. (2024).
Watervlakken 2024: Polygonenkaart van stilstaand water in
Vlaanderen. Uitgave 2024.
Rapporten van het Instituut voor Natuur- en Bosonderzoek 2024
(52). Instituut voor Natuur en Bosonderzoek, Brussel.
\doi{10.21436/inbor.114075267}.
\item De Saeger S., De Bruyn A., Dhaluin P., Erens R., Guelinckx G.,
Hennebel D., Jacobs I., Kumpen M., Van Oost F., Cool R., Spanhove T.,
Leyssen A., Oosterlynck P., Van Dam G. & Wils C. (red.) (2025). Biologische
Waarderingskaart en Natura 2000 Habitatkaart, uitgave 2025. Rapporten van het
Instituut voor Natuur- en Bosonderzoek 2025 (38). Instituut voor Natuur- en
Bosonderzoek, Brussel.
\doi{10.21436/inbor.129502912}.
}
}
\seealso{
Other functions involved in processing the 'habitatmap' data source: 
\code{\link{read_habitatmap}()},
\code{\link{read_habitatmap_stdized}()},
\code{\link{read_habitatmap_terr}()},
\code{\link{read_watersurfaces_refpoints}()}

Other functions involved in processing the 'watersurfaces' data source: 
\code{\link{read_watersurfaces}()},
\code{\link{read_watersurfaces_refpoints}()}
}
\concept{functions involved in processing the 'habitatmap' data source}
\concept{functions involved in processing the 'watersurfaces' data source}
