% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_habitatdata.R
\name{read_watersurfaces_hab}
\alias{read_watersurfaces_hab}
\title{Return the data source \code{watersurfaces_hab} as a list of two
objects}
\usage{
read_watersurfaces_hab(path = fileman_up("n2khab_data"),
  file = "20_processed/watersurfaces_hab/watersurfaces_hab.gpkg",
  interpreted = FALSE)
}
\arguments{
\item{path}{Location of the file.
Considering the default value of the \code{file} argument, this should be
the location of the folder '\strong{\code{n2khab_data}}'.
By default, the first \code{n2khab_data} folder is used that is found when
sequentially climbing up 0 to 10 levels in the file system hierarchy,
starting from the working directory.}

\item{file}{The filename of the data source.
May include a path prefix.
The default follows the data management advice in the
\href{../doc/v020_datastorage.html}{vignette} on data storage.}

\item{interpreted}{If \code{TRUE}, the interpreted subtype is provided when the subtype is missing. This only
applies to type 3130. When the subtype is missing for 3130, we interpret it as 3130_aom.}
}
\value{
A list of two objects:
  \itemize{
  \item \code{watersurfaces_hab_polygons}: an sf object of standing water habitat and RIB polygons with two attribute variables
  \itemize{
    \item \code{polygon_id}
    \item \code{polygon_id_ws}: id for the polygon in the \code{watersurface map}
    \item \code{polygon_id_habitatmap}: id's of all overlapping polygons of \code{habitatmap_stdized} that
    contain standing water habitat. The different id's are separated by '+'.
    \item \code{description_orig}: descriptions of all overlapping polygons of \code{habitatmap_stdized} that
    contain standing water habitat. The different descriptions are separated by '+'.}
  }
  \itemize{
  \item \code{habitatmap_patches}: a tibble with following variables
  \itemize{
    \item \code{polygon_id}
    \item \code{patch_id}
    \item \code{code_orig}: original vegetation code in raw \code{habitatmap}.
    \item \code{certain}: \code{TRUE} when vegetation type is certain and
     \code{FALSE} when vegetation type is uncertain.
    \item \code{type}: habitat or RIB type listed in \code{\link{types}}.
    }
    }
}
\description{
\code{read_watersurfaces_hab} returns the data source \code{watersurfaces_hab} as a list of two objects:
\itemize{
  \item \code{watersurfaces_hab_polygons}: an sf object with all polygons
  that contain standing water habitat or standing water Regionally
  Important Biotope (RIB).
  \item \code{watersurfaces_hab_patches}: a tibble with information on the
  standing water habitat and RIB patches (HAB1, HAB2,..., HAB5) that occur within
  the \code{watersurfaces_hab_polygons}, each row corresponding with one
  patch.
  }
}
\details{
The data source \code{watersurfaces_hab} is a combination of \code{habitatmap_stdized} (see
\code{\link{read_habitatmap_stdized}}) and the \href{https://http://www.geopunt.be/catalogus/datasetfolder/10e87ad3-8235-40e0-8269-42c3c96a884d}{watersurface map of Flanders}.
It contains all aquatic standing water habitat types and RIB in Flanders.

The data source \code{watersurfaces_hab} is a GeoPackage that
contains:
\itemize{
  \item \code{watersurfaces_hab_polygons}: a spatial layer with all polygons that contain standing water habitat
  or RIB type listed in \code{\link{types}}.
  \item \code{watersurfaces_hab_patches}: a table in which every row corresponds with one patch.
  }

The polygons with 2190_a habitat (dune slack ponds) are generated by selecting all watersurface polygons that
overlap with dune habitat polygons (21xx) of the processed habitat map.

For each of the other standing water habitat types (31xx, 7220 and rbbah) we select the watersurface polygons that
overlap with the selected habitat type polygons of the processed habitat map. We also select polygons of the
processed habitat map containing standing water habitat but not overlapping with any watersurface polygon of the
watersurface map.

The R-code for creating the \code{habitatmap_stdized} data source can be found in the \href{https://github.com/inbo/n2khab-preprocessing}{n2khab-preprocessing}
repository.
}
\examples{
\dontrun{
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'watersurfaces_hab'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

r <- read_watersurfaces_hab()
r_polygons <- r$watersurfaces_hab_polygons
r_patches <- r$watersurfaces_hab_patches
}

}
\seealso{
Other functions involved in processing the \code{habitatmap} data source: \code{\link{read_habitatmap_stdized}},
  \code{\link{read_habitatmap}}
}
\concept{functions involved in processing the \code{habitatmap} data source}
