% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata.R
\name{types}
\alias{types}
\title{Documentation of included data source 'types'}
\format{A vc-formatted data source. As such, it corresponds to
a dataframe with 7 variables:
\describe{
  \item{type}{Code of the type, as a factor.
  This is the ID for use in diverse workflows and datasets.
  Names in multiple languages are to be found in \code{\link{namelist}}.
  Only \emph{currently active} codes are kept in this list, in order to avoid
  confusion (this especially relates to habitat subtypes and RIBs).
  Contains no duplicates!}
  \item{typelevel}{A factor that labels the type as
   either \code{"main_type"} or \code{"subtype"}.}
  \item{main_type}{The main type that corresponds with \code{type},
  as a factor.
  Each type is either a subtype of a main type, or is a main type itself.
  This is indicated by \code{typelevel}.}
  \item{typeclass}{A code explained by \code{\link{namelist}},
  corresponding to the typeclass.
  Is a factor.}
  \item{tag_1}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.}
  \item{tag_2}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.}
  \item{tag_3}{Optional tag, e.g. a categorization ID explained
  by \code{\link{namelist}}.} }}
\source{
\href{https://docs.google.com/spreadsheets/d/1dK0S1Tt3RlVEh4WrNF5N_-hn0OXiTUVOvsL2JRhSl78}{This googlesheet}.
Currently, the googlesheet and the data source are both kept up-to-date.
}
\description{
'types' is a data source in the
\href{https://inbo.github.io/git2rdata/index.html}{vc-format} which provides
a checklist of types, represented by their \strong{current} codes, together
with several attributes.
A 'type' refers to either a (main) habitat type, a
habitat subtype or a regionally important biotope (RIB).
The codes of types, typeclasses and tags are explained in the
data source \code{\link{namelist}} (which can accommodate multiple
languages).
}
\section{Typical way of loading}{


  \code{read_types()}

  \code{read_types(lang = "nl")}
}

\section{Corresponding datafiles in the installed package}{


  \code{textdata/types.tsv}

  \code{textdata/types.yml}
}

\seealso{
\code{\link{read_types}}

Other n2khab-referencelists: \code{\link{env_pressures}},
  \code{\link{namelist}}, \code{\link{scheme_types}},
  \code{\link{schemes}}
}
\concept{n2khab-referencelists}
